+++ Parse/content.py
@@ -1,0 +1,2 @@
+class ContentParser:
+	pass
