+++ input.py
@@ -1,2088 +1,1 @@
-links_list = ['http://spiewnikreligijny.pl/teksty/4777,22', 'http://spiewnikreligijny.pl/teksty/3406,888',
-              'http://spiewnikreligijny.pl/teksty/298,a_czemuz_moj_jezus',
-              'http://spiewnikreligijny.pl/teksty/58,a_gu_gu', 'http://spiewnikreligijny.pl/teksty/2391,a_ja_czekam',
-              'http://spiewnikreligijny.pl/teksty/643,a_mysmy_sie_chlubic_powinni',
-              'http://spiewnikreligijny.pl/teksty/3393,a_slowo_cialem_sie_stalo',
-              'http://spiewnikreligijny.pl/teksty/303,a_teraz_wiem',
-              'http://spiewnikreligijny.pl/teksty/299,a_wczora_z_wieczora',
-              'http://spiewnikreligijny.pl/teksty/3980,abba_(jestes_radoscia_moja)',
-              'http://spiewnikreligijny.pl/teksty/1911,abba_naleze_do_ciebie',
-              'http://spiewnikreligijny.pl/teksty/626,abba_ojcze',
-              'http://spiewnikreligijny.pl/teksty/627,abba_ojcze_wolam',
-              'http://spiewnikreligijny.pl/teksty/1732,abba_abba_tatusiu',
-              'http://spiewnikreligijny.pl/teksty/71,abecadlo_z_nieba_spadlo',
-              'http://spiewnikreligijny.pl/teksty/4262,abraham',
-              'http://spiewnikreligijny.pl/teksty/3032,absolutnie_nic_nie_udalo_by_sie_nam_bez_milosci',
-              'http://spiewnikreligijny.pl/teksty/3192,aby_byli_jedno_musza_wydac_zycie',
-              'http://spiewnikreligijny.pl/teksty/628,aby_byli_jedno_zeby_swiat_zobaczyl',
-              'http://spiewnikreligijny.pl/teksty/2572,aby_uwielbic_jezusa',
-              'http://spiewnikreligijny.pl/teksty/629,abysmy_byli_jedno_blagamy_cie_ojcze',
-              'http://spiewnikreligijny.pl/teksty/630,abysmy_byli_jedno_podajmy_sobie_rece',
-              'http://spiewnikreligijny.pl/teksty/631,abysmy_mogli', 'http://spiewnikreligijny.pl/teksty/632,ach_kim',
-              'http://spiewnikreligijny.pl/teksty/296,ach_ubogi_zlobie',
-              'http://spiewnikreligijny.pl/teksty/297,ach_witajze_pozadana',
-              'http://spiewnikreligijny.pl/teksty/101,ach_moj_jezu',
-              'http://spiewnikreligijny.pl/teksty/102,ach_moj_jezu_ukochany',
-              'http://spiewnikreligijny.pl/teksty/304,adeste_fideles', 'http://spiewnikreligijny.pl/teksty/3823,adonai',
-              'http://spiewnikreligijny.pl/teksty/3399,adonai_elohenu',
-              'http://spiewnikreligijny.pl/teksty/3335,adonaj',
-              'http://spiewnikreligijny.pl/teksty/634,adoramus_te_christe',
-              'http://spiewnikreligijny.pl/teksty/633,adoramus_te_domine',
-              'http://spiewnikreligijny.pl/teksty/1064,adwentowa_lampka_lsni',
-              'http://spiewnikreligijny.pl/teksty/4277,agnus_dei',
-              'http://spiewnikreligijny.pl/teksty/4909,agnus_dei_(antyimaginacja)',
-              'http://spiewnikreligijny.pl/teksty/484,akatyst', 'http://spiewnikreligijny.pl/teksty/996,akt_pokutny',
-              'http://spiewnikreligijny.pl/teksty/635,akt_skruchy', 'http://spiewnikreligijny.pl/teksty/3270,akt_wiary',
-              'http://spiewnikreligijny.pl/teksty/4874,albo_albo',
-              'http://spiewnikreligijny.pl/teksty/637,albowiem_tak_bog_umilowal_swiat',
-              'http://spiewnikreligijny.pl/teksty/555,ale_ty_palisz_papierosy',
-              'http://spiewnikreligijny.pl/teksty/3227,alleluja', 'http://spiewnikreligijny.pl/teksty/4279,alleluja',
-              'http://spiewnikreligijny.pl/teksty/3445,alleluja',
-              'http://spiewnikreligijny.pl/teksty/4037,alleluja_(przed_ewangelia)',
-              'http://spiewnikreligijny.pl/teksty/3258,alleluja_alle_alleluja',
-              'http://spiewnikreligijny.pl/teksty/1,alleluja_bija_dzwony',
-              'http://spiewnikreligijny.pl/teksty/2286,alleluja_chwalcie_pana_u_swiatyni_bram',
-              'http://spiewnikreligijny.pl/teksty/2972,alleluja_glory_alleluja',
-              'http://spiewnikreligijny.pl/teksty/79,alleluja_i_hosanna',
-              'http://spiewnikreligijny.pl/teksty/640,alleluja_milosc_twa',
-              'http://spiewnikreligijny.pl/teksty/2300,alleluja_on_jest_swiatlem',
-              'http://spiewnikreligijny.pl/teksty/2758,alleluja_panem_jest',
-              'http://spiewnikreligijny.pl/teksty/2831,alleluja_w_bogu_nadzieje_mam',
-              'http://spiewnikreligijny.pl/teksty/2,alleluja_jezus_zyje',
-              'http://spiewnikreligijny.pl/teksty/3,alleluja_zyw_juz_jest_smierci_zwyciezyciel',
-              'http://spiewnikreligijny.pl/teksty/638,alleluja_alleluja',
-              'http://spiewnikreligijny.pl/teksty/20,alleluja_jezus_zyje_on_co_za_nas_zycie_dal',
-              'http://spiewnikreligijny.pl/teksty/641,alleluja_wznos_pod_niebo_glos',
-              'http://spiewnikreligijny.pl/teksty/589,amen', 'http://spiewnikreligijny.pl/teksty/1590,amen_amen',
-              'http://spiewnikreligijny.pl/teksty/642,amen_amen_alleluja',
-              'http://spiewnikreligijny.pl/teksty/3178,amen_amen_bog_jest_miloscia',
-              'http://spiewnikreligijny.pl/teksty/921,amen_jak_maryja',
-              'http://spiewnikreligijny.pl/teksty/4320,amen_jezus_moim_panem_jest',
-              'http://spiewnikreligijny.pl/teksty/2004,amen_amen_amen',
-              'http://spiewnikreligijny.pl/teksty/919,ani_brat_moj_ani_siostra',
-              'http://spiewnikreligijny.pl/teksty/3277,ani_oko_ani_ucho',
-              'http://spiewnikreligijny.pl/teksty/4263,aniele_bozy',
-              'http://spiewnikreligijny.pl/teksty/617,aniele_bozy',
-              'http://spiewnikreligijny.pl/teksty/4094,aniele_bozy_w_niebo_lec',
-              'http://spiewnikreligijny.pl/teksty/302,anieli_w_niebie_spiewaja',
-              'http://spiewnikreligijny.pl/teksty/4901,anielska_jesien',
-              'http://spiewnikreligijny.pl/teksty/4627,anielski_orszak',
-              'http://spiewnikreligijny.pl/teksty/4896,aniol_kwietniowy',
-              'http://spiewnikreligijny.pl/teksty/4118,aniol_panski__niesie_slowa',
-              'http://spiewnikreligijny.pl/teksty/2359,aniol_panski_(modlitwa)',
-              'http://spiewnikreligijny.pl/teksty/300,aniol_panski_otoczony_swiatloscia',
-              'http://spiewnikreligijny.pl/teksty/301,aniol_pasterzom_mowil',
-              'http://spiewnikreligijny.pl/teksty/1622,aniol_rzekl_maryi',
-              'http://spiewnikreligijny.pl/teksty/1744,aniol_stroz',
-              'http://spiewnikreligijny.pl/teksty/4897,aniol_stroz_bodyguard',
-              'http://spiewnikreligijny.pl/teksty/4891,aniol_stroz_dziadka',
-              'http://spiewnikreligijny.pl/teksty/4892,aniol_wielkiego_postu',
-              'http://spiewnikreligijny.pl/teksty/4894,aniol_zwiastowania',
-              'http://spiewnikreligijny.pl/teksty/644,aniolek',
-              'http://spiewnikreligijny.pl/teksty/2500,aniolek_najladniejszy',
-              'http://spiewnikreligijny.pl/teksty/306,aniolowie_biali',
-              'http://spiewnikreligijny.pl/teksty/4839,aniolowie_nocy',
-              'http://spiewnikreligijny.pl/teksty/305,aniolow_chory_i_mieszkancy_ziemi',
-              'http://spiewnikreligijny.pl/teksty/548,anioly_do_mnie_wysylaj',
-              'http://spiewnikreligijny.pl/teksty/3286,antyimaginacja',
-              'http://spiewnikreligijny.pl/teksty/645,apel_jasnogorski',
-              'http://spiewnikreligijny.pl/teksty/1623,apel_to__slowo',
-              'http://spiewnikreligijny.pl/teksty/32,archaniol_bozy_gabriel',
-              'http://spiewnikreligijny.pl/teksty/920,armia_pana',
-              'http://spiewnikreligijny.pl/teksty/1847,ave_maria_(gdy_klecze_przed_toba)',
-              'http://spiewnikreligijny.pl/teksty/1877,ave_maria_(ty_ktora_zawsze_spiewasz)',
-              'http://spiewnikreligijny.pl/teksty/646,ave_maria_gratia_plena',
-              'http://spiewnikreligijny.pl/teksty/647,ave_maryja_(gdy_dzien_sie_kryje)',
-              'http://spiewnikreligijny.pl/teksty/1963,ave_maryja_(juz_swiatlo_dzienne)',
-              'http://spiewnikreligijny.pl/teksty/3159,ave_maryja_(laski_pelna)',
-              'http://spiewnikreligijny.pl/teksty/1327,ave_maryja_(pozdrawiam_ciebie_matko)',
-              'http://spiewnikreligijny.pl/teksty/1863,ave_maryja_(z_toba_pan)',
-              'http://spiewnikreligijny.pl/teksty/1864,ave_maryja_z_fatimy',
-              'http://spiewnikreligijny.pl/teksty/2563,ave_ave_maryja',
-              'http://spiewnikreligijny.pl/teksty/922,az_do_wczoraj',
-              'http://spiewnikreligijny.pl/teksty/3946,az_ponad_morz_i_szczytow_gor',
-              'http://spiewnikreligijny.pl/teksty/4880,ballada',
-              'http://spiewnikreligijny.pl/teksty/652,ballada_ewangeliczna',
-              'http://spiewnikreligijny.pl/teksty/653,ballada_misyjna',
-              'http://spiewnikreligijny.pl/teksty/654,ballada_o_celniku_mateuszu',
-              'http://spiewnikreligijny.pl/teksty/3854,ballada_o_pieknej_pani',
-              'http://spiewnikreligijny.pl/teksty/1087,balonem_w_niebo',
-              'http://spiewnikreligijny.pl/teksty/103,barabasz',
-              'http://spiewnikreligijny.pl/teksty/3777,barabaszu_wiedz',
-              'http://spiewnikreligijny.pl/teksty/104,baranek_niesie_brzemie_win',
-              'http://spiewnikreligijny.pl/teksty/655,baranku_bozy',
-              'http://spiewnikreligijny.pl/teksty/3175,baranku_daj_mi_zmartwychwstanie',
-              'http://spiewnikreligijny.pl/teksty/4779,bardziej_niz_ja', 'http://spiewnikreligijny.pl/teksty/656,barka',
-              'http://spiewnikreligijny.pl/teksty/3396,baruch_ata_adonai',
-              'http://spiewnikreligijny.pl/teksty/657,baruch_haszszem',
-              'http://spiewnikreligijny.pl/teksty/658,badz_blogoslawiona',
-              'http://spiewnikreligijny.pl/teksty/1241,badz_blogoslawiony',
-              'http://spiewnikreligijny.pl/teksty/659,badz_krolowa',
-              'http://spiewnikreligijny.pl/teksty/105,badz_mi_litosciw_boze_nieskonczony',
-              'http://spiewnikreligijny.pl/teksty/2820,badz_pozdrowiona',
-              'http://spiewnikreligijny.pl/teksty/565,badz_pozdrowiona_krolowo',
-              'http://spiewnikreligijny.pl/teksty/660,badz_pozdrowiona_laski_pelna',
-              'http://spiewnikreligijny.pl/teksty/662,badz_pozdrowiony_gosciu_nasz',
-              'http://spiewnikreligijny.pl/teksty/664,badz_swiatloscia_dla_swojego_ludu',
-              'http://spiewnikreligijny.pl/teksty/586,badz_wola_twoja',
-              'http://spiewnikreligijny.pl/teksty/923,badz_z_nami_w_kontakcie',
-              'http://spiewnikreligijny.pl/teksty/663,badzze_pozdrowiona',
-              'http://spiewnikreligijny.pl/teksty/572,benedictus',
-              'http://spiewnikreligijny.pl/teksty/4713,betlejem_jak_iskierka',
-              'http://spiewnikreligijny.pl/teksty/315,betlejem_swiete_miasteczko_wslawione',
-              'http://spiewnikreligijny.pl/teksty/4840,betlejemski_znaczek',
-              'http://spiewnikreligijny.pl/teksty/2986,bezbronny_jak_motyl',
-              'http://spiewnikreligijny.pl/teksty/4781,bezpiecznie',
-              'http://spiewnikreligijny.pl/teksty/4581,bede_chwalil_pana_na_wieki',
-              'http://spiewnikreligijny.pl/teksty/1745,bede_cie_slawil',
-              'http://spiewnikreligijny.pl/teksty/924,bede_mowil',
-              'http://spiewnikreligijny.pl/teksty/925,bede_pana_czcil',
-              'http://spiewnikreligijny.pl/teksty/1845,bede_spiewal_ci',
-              'http://spiewnikreligijny.pl/teksty/1914,bede_spiewal_na_czesc_pana',
-              'http://spiewnikreligijny.pl/teksty/927,bede_spiewal_na_czesc_pana',
-              'http://spiewnikreligijny.pl/teksty/926,bede_spiewal_tobie',
-              'http://spiewnikreligijny.pl/teksty/2207,bede_tanczyl_przed_twym_tronem__',
-              'http://spiewnikreligijny.pl/teksty/1331,bede_z_toba_szedl',
-              'http://spiewnikreligijny.pl/teksty/3963,bedziecie_moimi_swiadkami',
-              'http://spiewnikreligijny.pl/teksty/3775,bedziesz_milowal_boga',
-              'http://spiewnikreligijny.pl/teksty/3405,biada_biada',
-              'http://spiewnikreligijny.pl/teksty/4620,biala_gwiazdka',
-              'http://spiewnikreligijny.pl/teksty/4375,biala_gwiazdka',
-              'http://spiewnikreligijny.pl/teksty/3304,biala_koszula',
-              'http://spiewnikreligijny.pl/teksty/2454,biala_ziemia',
-              'http://spiewnikreligijny.pl/teksty/3492,biale_swieta', 'http://spiewnikreligijny.pl/teksty/1840,bicz',
-              'http://spiewnikreligijny.pl/teksty/3020,biedny_kto_ciebie',
-              'http://spiewnikreligijny.pl/teksty/4072,bili_go_trzcina', 'http://spiewnikreligijny.pl/teksty/4863,bjm',
-              'http://spiewnikreligijny.pl/teksty/665,bliska_przyszlosc',
-              'http://spiewnikreligijny.pl/teksty/106,bliskie_jest_krolestwo_boze',
-              'http://spiewnikreligijny.pl/teksty/1964,blisko_jestes_panie',
-              'http://spiewnikreligijny.pl/teksty/872,blisko_blisko',
-              'http://spiewnikreligijny.pl/teksty/108,blogoslawiony_komu_odpuszczone_sa_winy',
-              'http://spiewnikreligijny.pl/teksty/1329,blogoslaw_matko',
-              'http://spiewnikreligijny.pl/teksty/3181,blogoslaw_pana_izraelu',
-              'http://spiewnikreligijny.pl/teksty/3591,blogoslaw_panie_nas',
-              'http://spiewnikreligijny.pl/teksty/485,blogoslawcie_pana',
-              'http://spiewnikreligijny.pl/teksty/4866,blogoslawieni',
-              'http://spiewnikreligijny.pl/teksty/1665,blogoslawienstwo_darow',
-              'http://spiewnikreligijny.pl/teksty/1683,blogoslawie_cie',
-              'http://spiewnikreligijny.pl/teksty/1733,blogoslawie_pana',
-              'http://spiewnikreligijny.pl/teksty/666,blogoslawiona',
-              'http://spiewnikreligijny.pl/teksty/2996,blogoslawiona_wina_',
-              'http://spiewnikreligijny.pl/teksty/4026,blogoslawione_serca_',
-              'http://spiewnikreligijny.pl/teksty/2596,blogoslawiony',
-              'http://spiewnikreligijny.pl/teksty/667,bo_gory_moga_ustapic',
-              'http://spiewnikreligijny.pl/teksty/1153,bo_jak_smierc',
-              'http://spiewnikreligijny.pl/teksty/2768,bo_milosierny_jest_pan',
-              'http://spiewnikreligijny.pl/teksty/668,bo_nikt_nie_ma_z_nas',
-              'http://spiewnikreligijny.pl/teksty/3050,bo_od_switu',
-              'http://spiewnikreligijny.pl/teksty/1698,bo_swiat_to_my',
-              'http://spiewnikreligijny.pl/teksty/3117,bo_wielki',
-              'http://spiewnikreligijny.pl/teksty/669,bo_zakrolowal_pan',
-              'http://spiewnikreligijny.pl/teksty/3671,bogu_wyspiewam',
-              'http://spiewnikreligijny.pl/teksty/486,bogurodzica',
-              'http://spiewnikreligijny.pl/teksty/4787,bohater_(psalm_45)',
-              'http://spiewnikreligijny.pl/teksty/3058,bolesna_matko',
-              'http://spiewnikreligijny.pl/teksty/314,boscy_poslowie_o_swieci_anieli',
-              'http://spiewnikreligijny.pl/teksty/4022,boski_klejnot',
-              'http://spiewnikreligijny.pl/teksty/4784,bossa_nova_piesn',
-              'http://spiewnikreligijny.pl/teksty/1077,bosy_pastuszek',
-              'http://spiewnikreligijny.pl/teksty/1746,boza_radosc_jak_rzeka',
-              'http://spiewnikreligijny.pl/teksty/4125,boza_milosc_w_sercu_swym_dzisiaj_mam',
-              'http://spiewnikreligijny.pl/teksty/875,boze_cos_polske',
-              'http://spiewnikreligijny.pl/teksty/316,boze_dziecie_witaj_nam',
-              'http://spiewnikreligijny.pl/teksty/109,boze_litosciw_zmiluj_sie_nade_mna',
-              'http://spiewnikreligijny.pl/teksty/3196,boze_moj_boze',
-              'http://spiewnikreligijny.pl/teksty/2733,boze_moj_boze_szukam_ciebie',
-              'http://spiewnikreligijny.pl/teksty/2838,boze_moj_boze_moj',
-              'http://spiewnikreligijny.pl/teksty/1074,boze_narodzenie',
-              'http://spiewnikreligijny.pl/teksty/3319,boze_pelen_w_niebie_chwaly',
-              'http://spiewnikreligijny.pl/teksty/1332,boze_powolales_mnie',
-              'http://spiewnikreligijny.pl/teksty/2769,boze_twa_laska_nad_nami_jest',
-              'http://spiewnikreligijny.pl/teksty/107,boze_w_dobroci_nigdy_nie_przebrany',
-              'http://spiewnikreligijny.pl/teksty/33,boze_wieczny_boze_zywy',
-              'http://spiewnikreligijny.pl/teksty/4545,boze_wszechmogacy',
-              'http://spiewnikreligijny.pl/teksty/3229,bozy_baranku',
-              'http://spiewnikreligijny.pl/teksty/4358,bog_adonai', 'http://spiewnikreligijny.pl/teksty/487,bog_i_czas',
-              'http://spiewnikreligijny.pl/teksty/1328,bog_i_tylko_on',
-              'http://spiewnikreligijny.pl/teksty/672,bog_jest_miloscia',
-              'http://spiewnikreligijny.pl/teksty/671,bog_jest_miloscia',
-              'http://spiewnikreligijny.pl/teksty/2058,bog_jest_tu_bog_jest_tam',
-              'http://spiewnikreligijny.pl/teksty/673,bog_kocha_mnie_takiego_jakim_jestem',
-              'http://spiewnikreligijny.pl/teksty/2558,bog_miedzy_nami_zamieszkal',
-              'http://spiewnikreligijny.pl/teksty/962,bog_nasz_pan',
-              'http://spiewnikreligijny.pl/teksty/934,bog_nie_umarl_jezus_zyje',
-              'http://spiewnikreligijny.pl/teksty/3899,bog_obiecal',
-              'http://spiewnikreligijny.pl/teksty/2311,bog_sie_rodzi_pokoj_niesie',
-              'http://spiewnikreligijny.pl/teksty/3537,bog_sie_narodzil',
-              'http://spiewnikreligijny.pl/teksty/1129,bog_sie_narodzil_nam',
-              'http://spiewnikreligijny.pl/teksty/312,bog_sie_rodzi_gwiazda_wschodzi',
-              'http://spiewnikreligijny.pl/teksty/309,bog_sie_rodzi_moc_truchleje',
-              'http://spiewnikreligijny.pl/teksty/310,bog_sie_z_panny_narodzil',
-              'http://spiewnikreligijny.pl/teksty/3373,bog_tak_ukochal_swiat',
-              'http://spiewnikreligijny.pl/teksty/935,bog_tak_umilowal_swiat',
-              'http://spiewnikreligijny.pl/teksty/1056,bog_to_milosc',
-              'http://spiewnikreligijny.pl/teksty/3022,bog_umieral',
-              'http://spiewnikreligijny.pl/teksty/2741,bog_w_stajni',
-              'http://spiewnikreligijny.pl/teksty/1029,bog_wskazuje_cel',
-              'http://spiewnikreligijny.pl/teksty/911,bog_zamieszkal',
-              'http://spiewnikreligijny.pl/teksty/933,bog_moj_pan_daje__mi_moc',
-              'http://spiewnikreligijny.pl/teksty/311,bracia_patrzcie_jeno',
-              'http://spiewnikreligijny.pl/teksty/1333,braciszkowie_skrzydlaci',
-              'http://spiewnikreligijny.pl/teksty/3908,brzmi_radosne_alleluja',
-              'http://spiewnikreligijny.pl/teksty/313,brzmia_anielskie_cudne_pienia',
-              'http://spiewnikreligijny.pl/teksty/3251,by_byc_jak_jezus',
-              'http://spiewnikreligijny.pl/teksty/3403,by_the_rivers_of_babylon',
-              'http://spiewnikreligijny.pl/teksty/110,byc_blizej_ciebie_chce',
-              'http://spiewnikreligijny.pl/teksty/1335,byl_dzien',
-              'http://spiewnikreligijny.pl/teksty/674,byla_cicha_i_piekna_jak_wiosna',
-              'http://spiewnikreligijny.pl/teksty/2825,bylbym_niczym',
-              'http://spiewnikreligijny.pl/teksty/1012,bywaja_chwile',
-              'http://spiewnikreligijny.pl/teksty/1340,caly_do_ciebie_naleze',
-              'http://spiewnikreligijny.pl/teksty/675,caly_swiat_niech_spiewa',
-              'http://spiewnikreligijny.pl/teksty/549,caly_twoj',
-              'http://spiewnikreligijny.pl/teksty/1717,canticorum_iubilo',
-              'http://spiewnikreligijny.pl/teksty/4746,centrala_(brygada_kryzys)',
-              'http://spiewnikreligijny.pl/teksty/488,chcemy_spozyc_pismo_swiete',
-              'http://spiewnikreligijny.pl/teksty/3155,chce_byc_twoim_ziarnem',
-              'http://spiewnikreligijny.pl/teksty/3198,chce_cie_uwielbiac',
-              'http://spiewnikreligijny.pl/teksty/913,chce_do_ciebie_podobnym_byc',
-              'http://spiewnikreligijny.pl/teksty/676,chce_przestapic_jego_prog',
-              'http://spiewnikreligijny.pl/teksty/2682,chce_tak_jak_on',
-              'http://spiewnikreligijny.pl/teksty/4051,chce_tylko_bys_byl',
-              'http://spiewnikreligijny.pl/teksty/873,chce_wywyzszac_imie_twe',
-              'http://spiewnikreligijny.pl/teksty/89,chcialem_dobrze',
-              'http://spiewnikreligijny.pl/teksty/94,chleb_z_nieba',
-              'http://spiewnikreligijny.pl/teksty/1242,chlebie_najcichszy',
-              'http://spiewnikreligijny.pl/teksty/3480,choc_tyle_zalu_w_nas',
-              'http://spiewnikreligijny.pl/teksty/1700,choc_wiatr_juz_zamknal_ksiege',
-              'http://spiewnikreligijny.pl/teksty/3068,chocby_u_boku_mego',
-              'http://spiewnikreligijny.pl/teksty/3858,chodziles_panie',
-              'http://spiewnikreligijny.pl/teksty/2489,choinek_blask_rozswietla_noc',
-              'http://spiewnikreligijny.pl/teksty/2503,choinka',
-              'http://spiewnikreligijny.pl/teksty/2719,chryste_ty_nas_pytales',
-              'http://spiewnikreligijny.pl/teksty/1338,chryste_tys_pokojem',
-              'http://spiewnikreligijny.pl/teksty/114,chrystus_cierpial_za_nas',
-              'http://spiewnikreligijny.pl/teksty/115,chrystus_pan_bozy_syn',
-              'http://spiewnikreligijny.pl/teksty/865,chrystus_pan_bozy_syn',
-              'http://spiewnikreligijny.pl/teksty/677,chrystus_pan_karmi_nas',
-              'http://spiewnikreligijny.pl/teksty/10,chrystus_pan_w_niebo_wstepuje',
-              'http://spiewnikreligijny.pl/teksty/30,chrystus_pan_zmartwychwstal',
-              'http://spiewnikreligijny.pl/teksty/3874,chrystus_sie_narodzil',
-              'http://spiewnikreligijny.pl/teksty/1958,chrystus_spi',
-              'http://spiewnikreligijny.pl/teksty/678,chrystus_wodzem',
-              'http://spiewnikreligijny.pl/teksty/4,chrystus_zmartwychwstan_jest',
-              'http://spiewnikreligijny.pl/teksty/2178,chrzescijanin_tanczy',
-              'http://spiewnikreligijny.pl/teksty/2243,chrzescijanin_to_ja',
-              'http://spiewnikreligijny.pl/teksty/3194,chwalcie_go',
-              'http://spiewnikreligijny.pl/teksty/648,chwalcie_laki_umajone',
-              'http://spiewnikreligijny.pl/teksty/4792,chwalcie_pana_(psalm_136)',
-              'http://spiewnikreligijny.pl/teksty/2757,chwalcie_pana_narody',
-              'http://spiewnikreligijny.pl/teksty/1747,chwalcie_pana_niebios',
-              'http://spiewnikreligijny.pl/teksty/2755,chwalcie_pana_niebios_ii',
-              'http://spiewnikreligijny.pl/teksty/116,chwalebny_krzyz_zmartwychwstalego_pana',
-              'http://spiewnikreligijny.pl/teksty/679,chwale_ciebie_panie',
-              'http://spiewnikreligijny.pl/teksty/4032,chwale_cie_',
-              'http://spiewnikreligijny.pl/teksty/1843,chwale_cie_panie',
-              'http://spiewnikreligijny.pl/teksty/1341,chwale_mojego_boga',
-              'http://spiewnikreligijny.pl/teksty/1649,chwale_pana_wciaz',
-              'http://spiewnikreligijny.pl/teksty/3185,chwalic_chce_mego_pana',
-              'http://spiewnikreligijny.pl/teksty/3315,chwalimy_pana_dzis',
-              'http://spiewnikreligijny.pl/teksty/1853,chwalmy_go',
-              'http://spiewnikreligijny.pl/teksty/680,chwalmy_pana_w_rytmie_reggae',
-              'http://spiewnikreligijny.pl/teksty/4860,chwala', 'http://spiewnikreligijny.pl/teksty/2362,chwala',
-              'http://spiewnikreligijny.pl/teksty/3179,chwala_barankowi',
-              'http://spiewnikreligijny.pl/teksty/2956,chwala_bogu_ojcu',
-              'http://spiewnikreligijny.pl/teksty/1337,chwala_panu_i_czesc',
-              'http://spiewnikreligijny.pl/teksty/3375,chwala_tobie_krolu_wiekow',
-              'http://spiewnikreligijny.pl/teksty/3253,chwala_tobie_panie',
-              'http://spiewnikreligijny.pl/teksty/4553,chwaly_piesn',
-              'http://spiewnikreligijny.pl/teksty/4841,chwile_dwie_w_twoich_ramionach',
-              'http://spiewnikreligijny.pl/teksty/547,ci_co_tobie_zaufali',
-              'http://spiewnikreligijny.pl/teksty/2813,ci_co_zaufali_panu',
-              'http://spiewnikreligijny.pl/teksty/1748,ci_ktorzy_jahwe_ufaja',
-              'http://spiewnikreligijny.pl/teksty/620,cialo_i_dusza',
-              'http://spiewnikreligijny.pl/teksty/1183,ciagle_boje_sie',
-              'http://spiewnikreligijny.pl/teksty/3320,ciagle_szukam_jedynej_tej_',
-              'http://spiewnikreligijny.pl/teksty/3922,ciagle_zaczynam_od_nowa',
-              'http://spiewnikreligijny.pl/teksty/321,cicha_noc_swieta_noc_ii',
-              'http://spiewnikreligijny.pl/teksty/317,cicha_noc_swieta_noc',
-              'http://spiewnikreligijny.pl/teksty/319,cicha_polana',
-              'http://spiewnikreligijny.pl/teksty/324,cicho_wszedzie_spi_swiat_caly',
-              'http://spiewnikreligijny.pl/teksty/1749,cichy_zapada_zmrok',
-              'http://spiewnikreligijny.pl/teksty/3834,ciebie_boga_wyslawiamy',
-              'http://spiewnikreligijny.pl/teksty/1082,ciemna_nocy',
-              'http://spiewnikreligijny.pl/teksty/320,ciemne_niebo_swiat_otula',
-              'http://spiewnikreligijny.pl/teksty/3343,cien',
-              'http://spiewnikreligijny.pl/teksty/111,cierniami_uwienczona',
-              'http://spiewnikreligijny.pl/teksty/1805,cierpienie',
-              'http://spiewnikreligijny.pl/teksty/11,ciesz_sie_krolowo_anielska',
-              'http://spiewnikreligijny.pl/teksty/1088,cieszcie_sie',
-              'http://spiewnikreligijny.pl/teksty/3111,ciesze_sie',
-              'http://spiewnikreligijny.pl/teksty/323,cieszmy_sie_i_pod_niebiosy',
-              'http://spiewnikreligijny.pl/teksty/2419,cisza', 'http://spiewnikreligijny.pl/teksty/2401,co_za_radosc',
-              'http://spiewnikreligijny.pl/teksty/1750,co_za_radosc_uslyszec_dzis',
-              'http://spiewnikreligijny.pl/teksty/3623,co_zawinil',
-              'http://spiewnikreligijny.pl/teksty/112,co_znaczy_krzyz',
-              'http://spiewnikreligijny.pl/teksty/3075,co_zrobisz_gdy',
-              'http://spiewnikreligijny.pl/teksty/113,cokolwiek_w_swiecie_jest',
-              'http://spiewnikreligijny.pl/teksty/4239,com_przyrzekl_bogu',
-              'http://spiewnikreligijny.pl/teksty/3292,constans', 'http://spiewnikreligijny.pl/teksty/891,cos',
-              'http://spiewnikreligijny.pl/teksty/3268,cos_sie_swieci',
-              'http://spiewnikreligijny.pl/teksty/4055,corki_jerozolimskie',
-              'http://spiewnikreligijny.pl/teksty/322,coz_to_prosze_za_nowina',
-              'http://spiewnikreligijny.pl/teksty/3234,coz_ci_jezu_damy',
-              'http://spiewnikreligijny.pl/teksty/2685,crucem_tuam',
-              'http://spiewnikreligijny.pl/teksty/3435,cud_w_betlejem',
-              'http://spiewnikreligijny.pl/teksty/1052,cuda_cuda',
-              'http://spiewnikreligijny.pl/teksty/1208,cudowny_kwiecie',
-              'http://spiewnikreligijny.pl/teksty/3974,cyrenejczykiem_badz',
-              'http://spiewnikreligijny.pl/teksty/681,czarna_madonna',
-              'http://spiewnikreligijny.pl/teksty/4444,czarna_madonno_witaj_nam',
-              'http://spiewnikreligijny.pl/teksty/2995,czas_zbawienia',
-              'http://spiewnikreligijny.pl/teksty/1065,czcijmy_jezusa_czcijmy_go',
-              'http://spiewnikreligijny.pl/teksty/3495,czekam_na_ciebie_dobry_boze',
-              'http://spiewnikreligijny.pl/teksty/55,czekam_na_ciebie_jezu_moj_maly',
-              'http://spiewnikreligijny.pl/teksty/1336,czemu_dobry_panie',
-              'http://spiewnikreligijny.pl/teksty/4411,czemu_ja',
-              'http://spiewnikreligijny.pl/teksty/117,czlowieku_krzyza',
-              'http://spiewnikreligijny.pl/teksty/4823,cztery_jeden_zero_jan',
-              'http://spiewnikreligijny.pl/teksty/3564,cztery_lata',
-              'http://spiewnikreligijny.pl/teksty/536,czy_mnie_przyjmiesz',
-              'http://spiewnikreligijny.pl/teksty/3092,czy_wierzysz__(wyznanie_wiary)_',
-              'http://spiewnikreligijny.pl/teksty/682,czy_wy_wiecie_ze_jestescie_swiatynia',
-              'http://spiewnikreligijny.pl/teksty/3269,czyste',
-              'http://spiewnikreligijny.pl/teksty/3177,daj_mi_sile_twa',
-              'http://spiewnikreligijny.pl/teksty/99,daj_na_zgode',
-              'http://spiewnikreligijny.pl/teksty/4673,daj_panie_pokoj',
-              'http://spiewnikreligijny.pl/teksty/1066,daje_tobie_panie',
-              'http://spiewnikreligijny.pl/teksty/1795,dajmy_szanse', 'http://spiewnikreligijny.pl/teksty/4883,dam_dam',
-              'http://spiewnikreligijny.pl/teksty/948,dawid_na_pustyni',
-              'http://spiewnikreligijny.pl/teksty/3137,dawno_temu_jezus_umarl',
-              'http://spiewnikreligijny.pl/teksty/3413,debora',
-              'http://spiewnikreligijny.pl/teksty/68,dej_pon_bog_wieczor',
-              'http://spiewnikreligijny.pl/teksty/3287,dekalog_vii',
-              'http://spiewnikreligijny.pl/teksty/4877,deus_caritas_est',
-              'http://spiewnikreligijny.pl/teksty/489,dialog_zakochanych',
-              'http://spiewnikreligijny.pl/teksty/949,dio_del_celo',
-              'http://spiewnikreligijny.pl/teksty/4815,dla_ciebie',
-              'http://spiewnikreligijny.pl/teksty/2922,dla_ciebie_siedze_dla_ciebie_chodze',
-              'http://spiewnikreligijny.pl/teksty/3735,dla_ciebie_zyje_kazdy_kwiat',
-              'http://spiewnikreligijny.pl/teksty/325,dlaczego_dzisiaj_wsrod_nocy_dnieje',
-              'http://spiewnikreligijny.pl/teksty/2006,dlaczego_tak_trudna_droga',
-              'http://spiewnikreligijny.pl/teksty/307,dnia_jednego_o_polnocy',
-              'http://spiewnikreligijny.pl/teksty/326,do_betlejem_pelni_radosci',
-              'http://spiewnikreligijny.pl/teksty/683,do_bialego_pasterza',
-              'http://spiewnikreligijny.pl/teksty/1815,do_ciebie_panie',
-              'http://spiewnikreligijny.pl/teksty/121,do_ciebie_panie_pokornie_wolamy',
-              'http://spiewnikreligijny.pl/teksty/490,do_ducha_swietego_prosba_o_laske',
-              'http://spiewnikreligijny.pl/teksty/491,do_ducha_swietego_prosba_o_oswiecenie',
-              'http://spiewnikreligijny.pl/teksty/3096,do_gory_nogami',
-              'http://spiewnikreligijny.pl/teksty/2887,do_jej_stop',
-              'http://spiewnikreligijny.pl/teksty/3060,do_matki_bozej_wejherowskiej',
-              'http://spiewnikreligijny.pl/teksty/3388,do_mnie_wroc',
-              'http://spiewnikreligijny.pl/teksty/3631,do_oltarza_dzis_zblizam_sie_jezu',
-              'http://spiewnikreligijny.pl/teksty/4394,do_szopy_w_betlejem',
-              'http://spiewnikreligijny.pl/teksty/327,do_szopy_hej_pasterze',
-              'http://spiewnikreligijny.pl/teksty/2778,do_twojej_swiatyni',
-              'http://spiewnikreligijny.pl/teksty/3650,dobra_matka',
-              'http://spiewnikreligijny.pl/teksty/118,dobranoc_glowo_swieta',
-              'http://spiewnikreligijny.pl/teksty/4013,dobranoc_maryjo',
-              'http://spiewnikreligijny.pl/teksty/80,dobre_dobre_bardzo_dobre',
-              'http://spiewnikreligijny.pl/teksty/4739,dobry_dzien',
-              'http://spiewnikreligijny.pl/teksty/3959,dobry_pasterzu',
-              'http://spiewnikreligijny.pl/teksty/1949,dobrze_znasz',
-              'http://spiewnikreligijny.pl/teksty/3240,dobrze_ze_jestes',
-              'http://spiewnikreligijny.pl/teksty/1915,dokonala_sie_we_mnie_zmiana',
-              'http://spiewnikreligijny.pl/teksty/4878,dopoki',
-              'http://spiewnikreligijny.pl/teksty/3759,dopoki_bije_serce_me_',
-              'http://spiewnikreligijny.pl/teksty/898,dotknal_mnie_dzis_pan_',
-              'http://spiewnikreligijny.pl/teksty/868,dotknij_panie_moich_oczu_abym_przejrzal',
-              'http://spiewnikreligijny.pl/teksty/2998,dotyk_ognia', 'http://spiewnikreligijny.pl/teksty/3116,droga',
-              'http://spiewnikreligijny.pl/teksty/1813,droga', 'http://spiewnikreligijny.pl/teksty/4876,droga',
-              'http://spiewnikreligijny.pl/teksty/4526,droga_pasterza',
-              'http://spiewnikreligijny.pl/teksty/1094,droga_przez_las',
-              'http://spiewnikreligijny.pl/teksty/1606,druga_po_bogu', 'http://spiewnikreligijny.pl/teksty/4264,drzewo',
-              'http://spiewnikreligijny.pl/teksty/3278,drzewo', 'http://spiewnikreligijny.pl/teksty/3170,drzewo_krzyza',
-              'http://spiewnikreligijny.pl/teksty/119,drzewo_krzyza_surowe',
-              'http://spiewnikreligijny.pl/teksty/2966,duch_swiety_niech_jednoczy_nas',
-              'http://spiewnikreligijny.pl/teksty/4664,duchu_jahwe',
-              'http://spiewnikreligijny.pl/teksty/120,duchu_milosci',
-              'http://spiewnikreligijny.pl/teksty/874,duchu_milosci_wylewaj_sie',
-              'http://spiewnikreligijny.pl/teksty/1214,duchu_pocieszycielu',
-              'http://spiewnikreligijny.pl/teksty/684,duchu_swiety',
-              'http://spiewnikreligijny.pl/teksty/3818,duchu_swiety_cudownym_wiatrem_jestes',
-              'http://spiewnikreligijny.pl/teksty/3226,duchu_swiety_do_nas_przybadz',
-              'http://spiewnikreligijny.pl/teksty/4849,duchu_swiety_do_nas_przybadz',
-              'http://spiewnikreligijny.pl/teksty/3173,duchu_swiety_duszo_mej_duszy',
-              'http://spiewnikreligijny.pl/teksty/3460,duchu_swiety_napelnij_serce_miloscia',
-              'http://spiewnikreligijny.pl/teksty/3392,duchu_swiety_ogarnij_mnie',
-              'http://spiewnikreligijny.pl/teksty/83,duchu_swiety_pocieszaj_mnie',
-              'http://spiewnikreligijny.pl/teksty/4265,duchu_swiety_przybywaj________________________________',
-              'http://spiewnikreligijny.pl/teksty/685,duchu_swiety_przyjdz',
-              'http://spiewnikreligijny.pl/teksty/1868,duchu_swiety_przyjdz_napelnij_serce_me',
-              'http://spiewnikreligijny.pl/teksty/1816,duchu_swiety_wolam',
-              'http://spiewnikreligijny.pl/teksty/863,duchu_swiety_wolam_przyjdz',
-              'http://spiewnikreligijny.pl/teksty/3202,duchu_swiety_wypelnij_swa_swiatynie',
-              'http://spiewnikreligijny.pl/teksty/4570,duchu_zyjacego_boga',
-              'http://spiewnikreligijny.pl/teksty/3783,duszo_crystusowa',
-              'http://spiewnikreligijny.pl/teksty/4769,duszu_swoju',
-              'http://spiewnikreligijny.pl/teksty/4791,dwie_drogi',
-              'http://spiewnikreligijny.pl/teksty/3276,dwie_siostry',
-              'http://spiewnikreligijny.pl/teksty/1925,dym_jalowca',
-              'http://spiewnikreligijny.pl/teksty/1112,dymites_patrus',
-              'http://spiewnikreligijny.pl/teksty/4266,dzieci_boze',
-              'http://spiewnikreligijny.pl/teksty/329,dzieciatko_sie_narodzilo',
-              'http://spiewnikreligijny.pl/teksty/330,dziecina_mala_bog_stworca_swiata',
-              'http://spiewnikreligijny.pl/teksty/1817,dzieckiem_bozym_jestem',
-              'http://spiewnikreligijny.pl/teksty/4825,dziecko_z_cukru',
-              'http://spiewnikreligijny.pl/teksty/1089,dzielmy_sie_wiara_jak_chlebem',
-              'http://spiewnikreligijny.pl/teksty/1049,dzien_taki_w_roku',
-              'http://spiewnikreligijny.pl/teksty/3627,dzien_ten_mam',
-              'http://spiewnikreligijny.pl/teksty/331,dzien_to_jest_dzis_wesela',
-              'http://spiewnikreligijny.pl/teksty/4100,dziewico_saletynska',
-              'http://spiewnikreligijny.pl/teksty/4173,dziewiec_wiekow_wstecz',
-              'http://spiewnikreligijny.pl/teksty/2176,dziekczynienie_za_jana_pawla_ii',
-              'http://spiewnikreligijny.pl/teksty/122,dzieki_ci_boze_moj',
-              'http://spiewnikreligijny.pl/teksty/3398,dzieki_ci_panie_(w_obliczu_aniolow)',
-              'http://spiewnikreligijny.pl/teksty/2539,dzieki_ci_panie_za_cialo_twe_i_krew',
-              'http://spiewnikreligijny.pl/teksty/3296,dzieki_jezu',
-              'http://spiewnikreligijny.pl/teksty/3295,dzieki_o_panie_skladamy_dzieki',
-              'http://spiewnikreligijny.pl/teksty/1818,dzieki_za_twoj_krzyz',
-              'http://spiewnikreligijny.pl/teksty/1967,dziekowac_ci_panie_chce',
-              'http://spiewnikreligijny.pl/teksty/1679,dziekuje_ci_boze_',
-              'http://spiewnikreligijny.pl/teksty/2828,dziekujemy_ci_ojcze__nasz',
-              'http://spiewnikreligijny.pl/teksty/4765,dziekuje_ci',
-              'http://spiewnikreligijny.pl/teksty/1344,dziekuje_ci_maryjo',
-              'http://spiewnikreligijny.pl/teksty/1852,dziekuje_ci_ojcze',
-              'http://spiewnikreligijny.pl/teksty/4552,dziekuje_tobie_panie',
-              'http://spiewnikreligijny.pl/teksty/332,dzisiaj_w_betlejem',
-              'http://spiewnikreligijny.pl/teksty/12,dzis_chrystus_krol_wiecznej_chwaly',
-              'http://spiewnikreligijny.pl/teksty/3574,dzis_do_betlejem',
-              'http://spiewnikreligijny.pl/teksty/4719,dzis_jest_czas_by_oddac_bogu_chwale',
-              'http://spiewnikreligijny.pl/teksty/2716,dzis_kosciele',
-              'http://spiewnikreligijny.pl/teksty/4198,dzis_w_stajence',
-              'http://spiewnikreligijny.pl/teksty/1078,dzis_z_radoscia_wszyscy_spiewajmy',
-              'http://spiewnikreligijny.pl/teksty/2753,dzis_zanosze_na_twoj_oltarz_panie',
-              'http://spiewnikreligijny.pl/teksty/123,ecce_homo', 'http://spiewnikreligijny.pl/teksty/601,effatha',
-              'http://spiewnikreligijny.pl/teksty/4801,efraim', 'http://spiewnikreligijny.pl/teksty/4852,ejn_szeni',
-              'http://spiewnikreligijny.pl/teksty/2978,el_szaddaj', 'http://spiewnikreligijny.pl/teksty/3162,eli_eli',
-              'http://spiewnikreligijny.pl/teksty/4768,eli_eli_lamma_sabahtani',
-              'http://spiewnikreligijny.pl/teksty/688,emmanuel',
-              'http://spiewnikreligijny.pl/teksty/3412,emmanuel_(narod_kroczacy)',
-              'http://spiewnikreligijny.pl/teksty/689,entuzjazm', 'http://spiewnikreligijny.pl/teksty/3069,eucharystia',
-              'http://spiewnikreligijny.pl/teksty/690,exodus', 'http://spiewnikreligijny.pl/teksty/1661,fala',
-              'http://spiewnikreligijny.pl/teksty/691,fanfary',
-              'http://spiewnikreligijny.pl/teksty/692,fantastyczne_spiewanie',
-              'http://spiewnikreligijny.pl/teksty/3339,father',
-              'http://spiewnikreligijny.pl/teksty/4757,faustyna_przypomina',
-              'http://spiewnikreligijny.pl/teksty/693,flaga', 'http://spiewnikreligijny.pl/teksty/694,francesco',
-              'http://spiewnikreligijny.pl/teksty/695,franciszku_jakzes_ty_odgadl',
-              'http://spiewnikreligijny.pl/teksty/696,furtka', 'http://spiewnikreligijny.pl/teksty/4764,gadu_gadu',
-              'http://spiewnikreligijny.pl/teksty/1621,gaude_mater',
-              'http://spiewnikreligijny.pl/teksty/4741,gaz_na_ulicach',
-              'http://spiewnikreligijny.pl/teksty/1224,gdy_adwentowy_wieczor_nadchodzi',
-              'http://spiewnikreligijny.pl/teksty/697,gdy_bozy_duch_wypelnia_mnie',
-              'http://spiewnikreligijny.pl/teksty/1819,gdy_chrzescijanin_tanczy',
-              'http://spiewnikreligijny.pl/teksty/130,gdy_czern_nas_przytlacza_kamienna',
-              'http://spiewnikreligijny.pl/teksty/1959,gdy_gasnie_dzien',
-              'http://spiewnikreligijny.pl/teksty/698,gdy_ide_w_swiat',
-              'http://spiewnikreligijny.pl/teksty/4112,gdy_idziemy_po_przez_swiat_______________',
-              'http://spiewnikreligijny.pl/teksty/132,gdy_ja_sobie_to_rozwazam',
-              'http://spiewnikreligijny.pl/teksty/699,gdy_kiedys_pan',
-              'http://spiewnikreligijny.pl/teksty/1820,gdy_klecze_przed_toba',
-              'http://spiewnikreligijny.pl/teksty/1879,gdy_marze',
-              'http://spiewnikreligijny.pl/teksty/3642,gdy_mily_jezus_byl_w_betanii',
-              'http://spiewnikreligijny.pl/teksty/700,gdy_na_aniol_panski',
-              'http://spiewnikreligijny.pl/teksty/701,gdy_na_morzu_wielka_burza',
-              'http://spiewnikreligijny.pl/teksty/4107,gdy_nas_ogarnie_trwoga',
-              'http://spiewnikreligijny.pl/teksty/2906,gdy_pan_jezus_byl_malutki',
-              'http://spiewnikreligijny.pl/teksty/334,gdy_pierwsza_gwiazda_niebo_rozjasni',
-              'http://spiewnikreligijny.pl/teksty/4480,gdy_pierwsza_gwiazdka_zjawi_sie',
-              'http://spiewnikreligijny.pl/teksty/2182,gdy_po_wielkim_polowie',
-              'http://spiewnikreligijny.pl/teksty/1839,gdy_schodzimy_sie',
-              'http://spiewnikreligijny.pl/teksty/333,gdy_sie_chrystus_rodzi',
-              'http://spiewnikreligijny.pl/teksty/3912,gdy_smutny_dzien',
-              'http://spiewnikreligijny.pl/teksty/128,gdy_srebrnikow_garsc',
-              'http://spiewnikreligijny.pl/teksty/2657,gdy_szukasz_boga',
-              'http://spiewnikreligijny.pl/teksty/335,gdy_sliczna_panna',
-              'http://spiewnikreligijny.pl/teksty/4115,gdy_trwoga_nas_ogarnie',
-              'http://spiewnikreligijny.pl/teksty/131,gdy_widze_ciebie_krwia_oblanego',
-              'http://spiewnikreligijny.pl/teksty/1165,gdy_wszyscy_swieci',
-              'http://spiewnikreligijny.pl/teksty/4170,gdy_zapada_grudniowa_noc',
-              'http://spiewnikreligijny.pl/teksty/2566,gdyby_nie_gwiazda_z_betlejem',
-              'http://spiewnikreligijny.pl/teksty/3190,gdyby_pan_nie_byl_po_naszej_stronie',
-              'http://spiewnikreligijny.pl/teksty/702,gdyby_wiara_twa',
-              'http://spiewnikreligijny.pl/teksty/3382,gdybym_przechodzil_przez_ciemna_doline',
-              'http://spiewnikreligijny.pl/teksty/3180,gdybys_tu_byl',
-              'http://spiewnikreligijny.pl/teksty/125,gdzie_duszy_ojczyzna',
-              'http://spiewnikreligijny.pl/teksty/2722,gdzie_dwaj',
-              'http://spiewnikreligijny.pl/teksty/337,gdzie_milosc_wzajemna_i_dobroc',
-              'http://spiewnikreligijny.pl/teksty/599,gdzie_ona_jest',
-              'http://spiewnikreligijny.pl/teksty/353,gdzie_taki_malutki',
-              'http://spiewnikreligijny.pl/teksty/3395,genesis', 'http://spiewnikreligijny.pl/teksty/3334,get_up',
-              'http://spiewnikreligijny.pl/teksty/575,getsemani', 'http://spiewnikreligijny.pl/teksty/3522,gloria',
-              'http://spiewnikreligijny.pl/teksty/3256,gloria_in_excelsis_deo',
-              'http://spiewnikreligijny.pl/teksty/336,gloria_gloria_zawolajmy',
-              'http://spiewnikreligijny.pl/teksty/3016,glos_boga',
-              'http://spiewnikreligijny.pl/teksty/34,glos_wdzieczny_z_nieba_wychodzi',
-              'http://spiewnikreligijny.pl/teksty/4408,glosem_twoim',
-              'http://spiewnikreligijny.pl/teksty/4828,glowa_na_pol',
-              'http://spiewnikreligijny.pl/teksty/3676,godzien_jestes',
-              'http://spiewnikreligijny.pl/teksty/1395,godzien_o_godzien_jest_bog',
-              'http://spiewnikreligijny.pl/teksty/1216,godzinki_do_matki_bozej_nieustajacej_pomocy',
-              'http://spiewnikreligijny.pl/teksty/2653,godzinki_o_mece_panskiej',
-              'http://spiewnikreligijny.pl/teksty/703,godzinki_o_niepokalanym_poczeciu_nmp',
-              'http://spiewnikreligijny.pl/teksty/129,golgoto', 'http://spiewnikreligijny.pl/teksty/3345,good_religion',
-              'http://spiewnikreligijny.pl/teksty/3336,gore',
-              'http://spiewnikreligijny.pl/teksty/67,gore_gwiazda_jezusowi',
-              'http://spiewnikreligijny.pl/teksty/2486,goralska_koleda', 'http://spiewnikreligijny.pl/teksty/4737,gps',
-              'http://spiewnikreligijny.pl/teksty/4155,grajmy_panu',
-              'http://spiewnikreligijny.pl/teksty/3960,grajmy_panu_na_harfie',
-              'http://spiewnikreligijny.pl/teksty/3290,grobowiec',
-              'http://spiewnikreligijny.pl/teksty/4906,gruchnela_nowina',
-              'http://spiewnikreligijny.pl/teksty/35,grzechem_adama_ludzie_uwiklani',
-              'http://spiewnikreligijny.pl/teksty/1796,gwiazdeczka', 'http://spiewnikreligijny.pl/teksty/4190,gwiazdka',
-              'http://spiewnikreligijny.pl/teksty/1960,gwiazdo_sliczna',
-              'http://spiewnikreligijny.pl/teksty/398,gwiazdo_zlota_promienista',
-              'http://spiewnikreligijny.pl/teksty/1044,gwiazdy_tanczyly',
-              'http://spiewnikreligijny.pl/teksty/709,habit', 'http://spiewnikreligijny.pl/teksty/4766,haraqia',
-              'http://spiewnikreligijny.pl/teksty/3002,hava_nagila',
-              'http://spiewnikreligijny.pl/teksty/708,hej_jezu_krolem_tys',
-              'http://spiewnikreligijny.pl/teksty/4556,hej_jezusicku',
-              'http://spiewnikreligijny.pl/teksty/338,hej_koleda_koleda',
-              'http://spiewnikreligijny.pl/teksty/4907,hej_koledo',
-              'http://spiewnikreligijny.pl/teksty/1821,hej_pasterzu',
-              'http://spiewnikreligijny.pl/teksty/707,hej_pawle',
-              'http://spiewnikreligijny.pl/teksty/340,hej_w_dzien_narodzenia',
-              'http://spiewnikreligijny.pl/teksty/36,hejnal_wszyscy_zaspiewajmy',
-              'http://spiewnikreligijny.pl/teksty/1663,historia_(gabriel_zwiastowal)',
-              'http://spiewnikreligijny.pl/teksty/2251,ho_ho_hosanna', 'http://spiewnikreligijny.pl/teksty/69,hojka',
-              'http://spiewnikreligijny.pl/teksty/3418,homilia_paschalna_melitona_z_sardes',
-              'http://spiewnikreligijny.pl/teksty/4748,horror_(oddzial_zamkniety)',
-              'http://spiewnikreligijny.pl/teksty/1797,horyzont', 'http://spiewnikreligijny.pl/teksty/3368,hosanna',
-              'http://spiewnikreligijny.pl/teksty/1951,hosanna',
-              'http://spiewnikreligijny.pl/teksty/1682,hosanna_memu_bogu',
-              'http://spiewnikreligijny.pl/teksty/4776,hymn_2',
-              'http://spiewnikreligijny.pl/teksty/492,hymn_do_ducha_sw',
-              'http://spiewnikreligijny.pl/teksty/2664,hymn_do_ducha_swietego',
-              'http://spiewnikreligijny.pl/teksty/493,hymn_iii_tysiaclecia',
-              'http://spiewnikreligijny.pl/teksty/2918,hymn_krwiodawcow',
-              'http://spiewnikreligijny.pl/teksty/494,hymn_o_domu_bozym',
-              'http://spiewnikreligijny.pl/teksty/495,hymn_o_krzyzu',
-              'http://spiewnikreligijny.pl/teksty/496,hymn_o_milosci',
-              'http://spiewnikreligijny.pl/teksty/3187,hymn_o_milosci_ii',
-              'http://spiewnikreligijny.pl/teksty/3859,hymn_o_milosci_iii',
-              'http://spiewnikreligijny.pl/teksty/497,hymn_o_pieczeci',
-              'http://spiewnikreligijny.pl/teksty/498,hymn_o_synu_bozym',
-              'http://spiewnikreligijny.pl/teksty/2693,hymn_wielkanocny',
-              'http://spiewnikreligijny.pl/teksty/4762,i_bez_kropki', 'http://spiewnikreligijny.pl/teksty/3340,i_feel',
-              'http://spiewnikreligijny.pl/teksty/3473,i_imie_ojca_i_syna',
-              'http://spiewnikreligijny.pl/teksty/1822,i_to_jest_wlasnie_to',
-              'http://spiewnikreligijny.pl/teksty/499,ichtis', 'http://spiewnikreligijny.pl/teksty/3342,ida_ida',
-              'http://spiewnikreligijny.pl/teksty/3540,ida_trzej_krolowie',
-              'http://spiewnikreligijny.pl/teksty/4875,idealne_polaczenie',
-              'http://spiewnikreligijny.pl/teksty/1520,ide_do_ciebie_panie',
-              'http://spiewnikreligijny.pl/teksty/622,ide_do_spowiedzi',
-              'http://spiewnikreligijny.pl/teksty/3900,ide_panie_na_prog_twoich_oltarzy',
-              'http://spiewnikreligijny.pl/teksty/4820,ide_panie_biegne',
-              'http://spiewnikreligijny.pl/teksty/3927,ide_za_toba_panie',
-              'http://spiewnikreligijny.pl/teksty/711,idzie_jezus',
-              'http://spiewnikreligijny.pl/teksty/712,idzie_moj_pan',
-              'http://spiewnikreligijny.pl/teksty/714,idzie_nasza_karawana',
-              'http://spiewnikreligijny.pl/teksty/3337,idziemy_do_ciebie',
-              'http://spiewnikreligijny.pl/teksty/713,idziemy_razem',
-              'http://spiewnikreligijny.pl/teksty/4429,idziemy_za_toba_jezu',
-              'http://spiewnikreligijny.pl/teksty/4635,idziesz_przez_wieki',
-              'http://spiewnikreligijny.pl/teksty/522,idzcie_juz_na_ulice',
-              'http://spiewnikreligijny.pl/teksty/4703,idzcie_na_caly_swiat',
-              'http://spiewnikreligijny.pl/teksty/715,idzmy_tulmy_sie_jak_dziatki',
-              'http://spiewnikreligijny.pl/teksty/578,ii_piesn_slugi_panskiego',
-              'http://spiewnikreligijny.pl/teksty/716,imie_jego_jest_jezus',
-              'http://spiewnikreligijny.pl/teksty/543,introductio_laudet_te_orbis_terrarum',
-              'http://spiewnikreligijny.pl/teksty/717,iskra',
-              'http://spiewnikreligijny.pl/teksty/584,iv_piesn_slugi_panskiego',
-              'http://spiewnikreligijny.pl/teksty/1142,izajasz', 'http://spiewnikreligijny.pl/teksty/63,ja_jestem',
-              'http://spiewnikreligijny.pl/teksty/3305,ja_nie_dam_rady',
-              'http://spiewnikreligijny.pl/teksty/84,ja_nie_umieram', 'http://spiewnikreligijny.pl/teksty/2589,ja_pan',
-              'http://spiewnikreligijny.pl/teksty/4089,ja_pan_jestem_swiadom',
-              'http://spiewnikreligijny.pl/teksty/3214,ja_spie', 'http://spiewnikreligijny.pl/teksty/618,ja_tylko_tak',
-              'http://spiewnikreligijny.pl/teksty/1628,ja_wiem_w_kogo_ja_wierze_(wersja_i)',
-              'http://spiewnikreligijny.pl/teksty/3303,ja_wiem_ze_tak_bedzie',
-              'http://spiewnikreligijny.pl/teksty/4846,ja_wiem_ze_jestes',
-              'http://spiewnikreligijny.pl/teksty/1823,ja_wierze_ze_to_jezus',
-              'http://spiewnikreligijny.pl/teksty/3385,jahwe',
-              'http://spiewnikreligijny.pl/teksty/581,jahwe_tys_bogiem_mym',
-              'http://spiewnikreligijny.pl/teksty/3122,jak_brzask',
-              'http://spiewnikreligijny.pl/teksty/3421,jak_cenna_jest_dla_mnie',
-              'http://spiewnikreligijny.pl/teksty/719,jak_ci_dziekowac',
-              'http://spiewnikreligijny.pl/teksty/720,jak_cie_szukac',
-              'http://spiewnikreligijny.pl/teksty/140,jak_cie_wital_niegdys_lud',
-              'http://spiewnikreligijny.pl/teksty/721,jak_cie_zrozumiec',
-              'http://spiewnikreligijny.pl/teksty/2688,jak_cudowne_to_jest',
-              'http://spiewnikreligijny.pl/teksty/1757,jak_dawid',
-              'http://spiewnikreligijny.pl/teksty/1965,jak_dobrze_jest_dziekowac',
-              'http://spiewnikreligijny.pl/teksty/4843,jak_lania', 'http://spiewnikreligijny.pl/teksty/726,jak_lania',
-              'http://spiewnikreligijny.pl/teksty/2451,jak_maryjo_ty',
-              'http://spiewnikreligijny.pl/teksty/1824,jak_mi_dobrze',
-              'http://spiewnikreligijny.pl/teksty/4826,jak_ogien',
-              'http://spiewnikreligijny.pl/teksty/1825,jak_ozywczy_deszcz',
-              'http://spiewnikreligijny.pl/teksty/4474,jak_paciorki_rozanca',
-              'http://spiewnikreligijny.pl/teksty/3273,jak_rzeka',
-              'http://spiewnikreligijny.pl/teksty/650,jak_szczesliwa_polska_cala',
-              'http://spiewnikreligijny.pl/teksty/727,jak_trawa',
-              'http://spiewnikreligijny.pl/teksty/3437,jak_trzej_krolowie',
-              'http://spiewnikreligijny.pl/teksty/1083,jak_woda_lubi_cisze',
-              'http://spiewnikreligijny.pl/teksty/4468,jak_zywa_woda',
-              'http://spiewnikreligijny.pl/teksty/718,jaka_radosc', 'http://spiewnikreligijny.pl/teksty/2938,jaka_sila',
-              'http://spiewnikreligijny.pl/teksty/3539,jakas_swiatlosc',
-              'http://spiewnikreligijny.pl/teksty/342,jakaz_to_gwiazda',
-              'http://spiewnikreligijny.pl/teksty/157,jakbym_bardzo_chcial',
-              'http://spiewnikreligijny.pl/teksty/722,jaki_byl_ten_dzien',
-              'http://spiewnikreligijny.pl/teksty/3538,jakze_dzisiaj_spiewny_swiat',
-              'http://spiewnikreligijny.pl/teksty/141,jam_gosciem_na_tym_swiecie',
-              'http://spiewnikreligijny.pl/teksty/3247,jam_nie_godzien',
-              'http://spiewnikreligijny.pl/teksty/1928,jamnenska_koronka_do_milosierdzia_bozego_(lednica)',
-              'http://spiewnikreligijny.pl/teksty/500,jamnenska_piesn_wieczorna',
-              'http://spiewnikreligijny.pl/teksty/4280,janie_pawle_teraz_przyjdz',
-              'http://spiewnikreligijny.pl/teksty/945,jaselka', 'http://spiewnikreligijny.pl/teksty/343,jasna_koleda',
-              'http://spiewnikreligijny.pl/teksty/556,jasmin', 'http://spiewnikreligijny.pl/teksty/728,jeden_chleb',
-              'http://spiewnikreligijny.pl/teksty/1053,jeden_duch',
-              'http://spiewnikreligijny.pl/teksty/3338,jeden_jest',
-              'http://spiewnikreligijny.pl/teksty/1969,jeden_jest_tylko_pan',
-              'http://spiewnikreligijny.pl/teksty/866,jedna_dzis_tworzac_rodzine',
-              'http://spiewnikreligijny.pl/teksty/729,jednego_mamy_ojca',
-              'http://spiewnikreligijny.pl/teksty/966,jedno_jest_imie_ktore_chce_opiewac_',
-              'http://spiewnikreligijny.pl/teksty/564,jedyny_krol',
-              'http://spiewnikreligijny.pl/teksty/3557,jedyny_nasz_bog',
-              'http://spiewnikreligijny.pl/teksty/1826,jedyny_pan',
-              'http://spiewnikreligijny.pl/teksty/4114,jeruzalem_jeruzalem',
-              'http://spiewnikreligijny.pl/teksty/1972,jeruzalem_moje_czy_slyszysz_placz',
-              'http://spiewnikreligijny.pl/teksty/1235,jest_drabina_do_nieba',
-              'http://spiewnikreligijny.pl/teksty/2154,jest_jedno_cialo',
-              'http://spiewnikreligijny.pl/teksty/4793,jest_ktos',
-              'http://spiewnikreligijny.pl/teksty/730,jest_na_swiecie_milosc',
-              'http://spiewnikreligijny.pl/teksty/2874,jest_taka_sila',
-              'http://spiewnikreligijny.pl/teksty/1081,jest_taki_dzien',
-              'http://spiewnikreligijny.pl/teksty/4803,jest_takie_miejsce',
-              'http://spiewnikreligijny.pl/teksty/3855,jest_we_mnie_kraina',
-              'http://spiewnikreligijny.pl/teksty/1827,jestem_chlebem_zycia',
-              'http://spiewnikreligijny.pl/teksty/3282,jestem_dzieckiem_boga',
-              'http://spiewnikreligijny.pl/teksty/3148,jestem_kochany',
-              'http://spiewnikreligijny.pl/teksty/2367,jestem_twoj',
-              'http://spiewnikreligijny.pl/teksty/3415,jestem_z_toba',
-              'http://spiewnikreligijny.pl/teksty/2942,jestes_blisko',
-              'http://spiewnikreligijny.pl/teksty/1016,jestes_blisko_mnie',
-              'http://spiewnikreligijny.pl/teksty/1942,jestes_dobry',
-              'http://spiewnikreligijny.pl/teksty/4802,jestes_dzieckiem',
-              'http://spiewnikreligijny.pl/teksty/731,jestes_krolem',
-              'http://spiewnikreligijny.pl/teksty/3284,jestes_ktory_jestes',
-              'http://spiewnikreligijny.pl/teksty/1828,jestes_panie_winnym_krzewem',
-              'http://spiewnikreligijny.pl/teksty/3832,jestes_perla',
-              'http://spiewnikreligijny.pl/teksty/732,jestes_radoscia_mojego_zycia',
-              'http://spiewnikreligijny.pl/teksty/623,jestes_slowem',
-              'http://spiewnikreligijny.pl/teksty/541,jestes_swiatlem',
-              'http://spiewnikreligijny.pl/teksty/2985,jestes_tuz_obok_mnie',
-              'http://spiewnikreligijny.pl/teksty/1026,jestes_zyciem_mym',
-              'http://spiewnikreligijny.pl/teksty/2960,jestescie_przyjaciolmi',
-              'http://spiewnikreligijny.pl/teksty/1684,jestesmy_jednym_cialem_w_chrystusie',
-              'http://spiewnikreligijny.pl/teksty/4482,jestesmy_piekni',
-              'http://spiewnikreligijny.pl/teksty/3109,jesus_loves_us',
-              'http://spiewnikreligijny.pl/teksty/4641,jeszcze_sie_kiedys_rozsmuce',
-              'http://spiewnikreligijny.pl/teksty/2335,jesli_na_piasku_zbudujesz_swoj_dom',
-              'http://spiewnikreligijny.pl/teksty/2970,jesli_pan_domu_nie_zbuduje',
-              'http://spiewnikreligijny.pl/teksty/1829,jesli_radosc',
-              'http://spiewnikreligijny.pl/teksty/2307,jesli_sie_w_niebie_spotkamy_kochanie',
-              'http://spiewnikreligijny.pl/teksty/1506,jezu_chce_do_ciebie_nalezec',
-              'http://spiewnikreligijny.pl/teksty/142,jezu_chryste_panie_mily',
-              'http://spiewnikreligijny.pl/teksty/4036,jezu_chryste_panie_moj',
-              'http://spiewnikreligijny.pl/teksty/3118,jezu_drogi_tys_miloscia',
-              'http://spiewnikreligijny.pl/teksty/3161,jezu_jezu_wspomnij_na_mnie',
-              'http://spiewnikreligijny.pl/teksty/346,jezu_mile_dzieciatko',
-              'http://spiewnikreligijny.pl/teksty/3128,jezu_mistrzu',
-              'http://spiewnikreligijny.pl/teksty/1830,jezu_moj_jezu',
-              'http://spiewnikreligijny.pl/teksty/2575,jezu_moj_jezu_dzis_do_ciebie_mowic_chce',
-              'http://spiewnikreligijny.pl/teksty/2883,jezu_potrzebuje__cie',
-              'http://spiewnikreligijny.pl/teksty/869,jezu_tchnij_w_moje_serce',
-              'http://spiewnikreligijny.pl/teksty/4649,jezu_ty_jestes_bogiem',
-              'http://spiewnikreligijny.pl/teksty/2220,jezu_ty_masz_moc',
-              'http://spiewnikreligijny.pl/teksty/1695,jezu_tys_jest_swiatloscia_mej_duszy',
-              'http://spiewnikreligijny.pl/teksty/3765,jezu_ufam_tobie',
-              'http://spiewnikreligijny.pl/teksty/289,jezu_zostan_w_nas',
-              'http://spiewnikreligijny.pl/teksty/4832,jezu_ufam_tobie_(piesn_pielgrzyma_na_sdm_2016)',
-              'http://spiewnikreligijny.pl/teksty/501,jezus_chrystus',
-              'http://spiewnikreligijny.pl/teksty/1831,jezus_chrystus_jest_moim_zbawieniem',
-              'http://spiewnikreligijny.pl/teksty/733,jezus_chrystus_moim_panem_jest',
-              'http://spiewnikreligijny.pl/teksty/3367,jezus_chrystus_panem_jest',
-              'http://spiewnikreligijny.pl/teksty/1832,jezus_chrystus_to_panow_pan',
-              'http://spiewnikreligijny.pl/teksty/738,jezus_daje_nam_zbawienie',
-              'http://spiewnikreligijny.pl/teksty/3636,jezus_dal_maryje',
-              'http://spiewnikreligijny.pl/teksty/4483,jezus_jak_cudowny_to_pan',
-              'http://spiewnikreligijny.pl/teksty/2381,jezus_jest_krolem',
-              'http://spiewnikreligijny.pl/teksty/735,jezus_jest_mym_przyjacielem',
-              'http://spiewnikreligijny.pl/teksty/1833,jezus_jest_nasza_sila',
-              'http://spiewnikreligijny.pl/teksty/567,jezus_jest_panem',
-              'http://spiewnikreligijny.pl/teksty/1306,jezus_jest_panem_ii',
-              'http://spiewnikreligijny.pl/teksty/1834,jezus_jest_tu',
-              'http://spiewnikreligijny.pl/teksty/1755,jezus_jest_z_nami_tu',
-              'http://spiewnikreligijny.pl/teksty/4017,jezus_kocha_ciebie_dzis',
-              'http://spiewnikreligijny.pl/teksty/139,jezus_krolem_naszym_jest',
-              'http://spiewnikreligijny.pl/teksty/3446,jezus_krolem_naszym_jest',
-              'http://spiewnikreligijny.pl/teksty/3968,jezus_lezy_w_stajence',
-              'http://spiewnikreligijny.pl/teksty/345,jezus_malusienki',
-              'http://spiewnikreligijny.pl/teksty/2270,jezus_ma_miloscia',
-              'http://spiewnikreligijny.pl/teksty/1835,jezus_moim_panem',
-              'http://spiewnikreligijny.pl/teksty/4133,jezus_moj_pan',
-              'http://spiewnikreligijny.pl/teksty/1836,jezus_najwyzsze_imie',
-              'http://spiewnikreligijny.pl/teksty/737,jezus_o_poranku',
-              'http://spiewnikreligijny.pl/teksty/4296,jezus_objawione_serce_boga',
-              'http://spiewnikreligijny.pl/teksty/3948,jezus_pokonal_smierc',
-              'http://spiewnikreligijny.pl/teksty/1880,jezus_przez_zycie',
-              'http://spiewnikreligijny.pl/teksty/66,jezus_ratownik',
-              'http://spiewnikreligijny.pl/teksty/1884,jezus_sila_ma',
-              'http://spiewnikreligijny.pl/teksty/739,jezus_swoja_matke_pozostawil',
-              'http://spiewnikreligijny.pl/teksty/3383,jezus_wychwalajmy_go',
-              'http://spiewnikreligijny.pl/teksty/4844,jezus_zbawicielem',
-              'http://spiewnikreligijny.pl/teksty/740,jezus_zwyciezyl',
-              'http://spiewnikreligijny.pl/teksty/344,jezusa_narodzonego',
-              'http://spiewnikreligijny.pl/teksty/3171,jezusie_synu_dawida',
-              'http://spiewnikreligijny.pl/teksty/2712,jek_grzesznika', 'http://spiewnikreligijny.pl/teksty/4867,jjp',
-              'http://spiewnikreligijny.pl/teksty/2455,jozefie_stajenki_nie_szukaj',
-              'http://spiewnikreligijny.pl/teksty/2791,judasz',
-              'http://spiewnikreligijny.pl/teksty/349,judzka_kraine_noc_okryla',
-              'http://spiewnikreligijny.pl/teksty/143,juz_chrystus_zycie_zakonczyl',
-              'http://spiewnikreligijny.pl/teksty/144,juz_cie_zegnam_najmilszy_synu',
-              'http://spiewnikreligijny.pl/teksty/1243,juz_goscisz_jezu',
-              'http://spiewnikreligijny.pl/teksty/1159,juz_gwiazdeczka',
-              'http://spiewnikreligijny.pl/teksty/4218,juz_przyszedl_jezus_ukochany',
-              'http://spiewnikreligijny.pl/teksty/348,juz_sie_piesn_rozlega',
-              'http://spiewnikreligijny.pl/teksty/741,juz_teraz_we_mnie_kwitna_twe_ogrody',
-              'http://spiewnikreligijny.pl/teksty/31,juz_zbawiciel_jezus_zyje',
-              'http://spiewnikreligijny.pl/teksty/138,juzem_dosc_pracowal',
-              'http://spiewnikreligijny.pl/teksty/4511,kaczka_pstra_dziatki_ma',
-              'http://spiewnikreligijny.pl/teksty/742,kanony_z_taize',
-              'http://spiewnikreligijny.pl/teksty/2675,kantyk_symeona',
-              'http://spiewnikreligijny.pl/teksty/570,kantyk_trzech_mlodziencow',
-              'http://spiewnikreligijny.pl/teksty/4886,karolina',
-              'http://spiewnikreligijny.pl/teksty/3302,kazda_noc_kazdy_dzien',
-              'http://spiewnikreligijny.pl/teksty/1837,kazdego_dnia',
-              'http://spiewnikreligijny.pl/teksty/2368,kazdego_dnia_przy_tobie_trwac',
-              'http://spiewnikreligijny.pl/teksty/2803,kazdy_chce_byc_kochany',
-              'http://spiewnikreligijny.pl/teksty/1203,kazdy_dzien_jest_zwyciestwem',
-              'http://spiewnikreligijny.pl/teksty/3779,kazdy_dzien_upewnia_mnie_i',
-              'http://spiewnikreligijny.pl/teksty/3788,kazdy_dzien_upewnia_mnie_ii',
-              'http://spiewnikreligijny.pl/teksty/744,kazdy_moze_dzis',
-              'http://spiewnikreligijny.pl/teksty/1885,kazdy_wschod_slonca',
-              'http://spiewnikreligijny.pl/teksty/3301,kefas', 'http://spiewnikreligijny.pl/teksty/2392,kiedy',
-              'http://spiewnikreligijny.pl/teksty/4091,kiedy_bozy_duch',
-              'http://spiewnikreligijny.pl/teksty/4016,kiedy_cichnie_gwar',
-              'http://spiewnikreligijny.pl/teksty/3804,kiedy_czuje_sie_samotny',
-              'http://spiewnikreligijny.pl/teksty/3293,kiedy_dusza_spiewa',
-              'http://spiewnikreligijny.pl/teksty/4804,kiedy_mamy_problem',
-              'http://spiewnikreligijny.pl/teksty/1844,kiedy_on_zaczyna_dzialac',
-              'http://spiewnikreligijny.pl/teksty/154,kiedy_patrze_na_twoj_krzyz',
-              'http://spiewnikreligijny.pl/teksty/745,kiedy_ranne_wstaja_zorze',
-              'http://spiewnikreligijny.pl/teksty/355,kiedy_slysze_na_oblokach',
-              'http://spiewnikreligijny.pl/teksty/291,kiedy_w_jasna_spokojna_cicha_noc',
-              'http://spiewnikreligijny.pl/teksty/4528,kiedy_widze_twoja_postac',
-              'http://spiewnikreligijny.pl/teksty/4513,kiedym_ja_byl_ogrodniczkiem',
-              'http://spiewnikreligijny.pl/teksty/37,kiedys_powroci_pan',
-              'http://spiewnikreligijny.pl/teksty/2244,kiedys_wino_i_chleb',
-              'http://spiewnikreligijny.pl/teksty/354,kiej_owiecek',
-              'http://spiewnikreligijny.pl/teksty/587,kim_jest_ta',
-              'http://spiewnikreligijny.pl/teksty/624,kim_on_jest',
-              'http://spiewnikreligijny.pl/teksty/3151,kimze_jest_ta_(oblubienica)',
-              'http://spiewnikreligijny.pl/teksty/1886,klaszczmy_w_dlonie',
-              'http://spiewnikreligijny.pl/teksty/150,klaniam_sie_tobie',
-              'http://spiewnikreligijny.pl/teksty/1887,kobierzec_maryjny',
-              'http://spiewnikreligijny.pl/teksty/1903,kobieta_pochylona',
-              'http://spiewnikreligijny.pl/teksty/2190,kochaj_boga',
-              'http://spiewnikreligijny.pl/teksty/3632,kochaj_jezusa_tak_jak_ja',
-              'http://spiewnikreligijny.pl/teksty/3348,kochaj_mnie',
-              'http://spiewnikreligijny.pl/teksty/1114,kocham_ciebie_jezu',
-              'http://spiewnikreligijny.pl/teksty/2048,kocham_cie_jezu',
-              'http://spiewnikreligijny.pl/teksty/1504,kocham_cie_panie',
-              'http://spiewnikreligijny.pl/teksty/3400,kocham_cie_tak_samo',
-              'http://spiewnikreligijny.pl/teksty/3380,kocham_wiec_nie_musze_sie_bac',
-              'http://spiewnikreligijny.pl/teksty/1201,kochany_bracie_kochana_siostro_(takie_jest_prawo_milosci)',
-              'http://spiewnikreligijny.pl/teksty/2511,koleda_andrychowska',
-              'http://spiewnikreligijny.pl/teksty/4905,koleda_dla_aniolka',
-              'http://spiewnikreligijny.pl/teksty/1200,koleda_dla_nieobecnych',
-              'http://spiewnikreligijny.pl/teksty/352,koleda_dla_piotra',
-              'http://spiewnikreligijny.pl/teksty/4805,koleda_o_sercu_(opowiesc_wigilijna)',
-              'http://spiewnikreligijny.pl/teksty/2409,koleda_plynie_z_wysokosci',
-              'http://spiewnikreligijny.pl/teksty/3532,koleda_sniezka',
-              'http://spiewnikreligijny.pl/teksty/4714,koleda_wedrujacych_do_betlejem',
-              'http://spiewnikreligijny.pl/teksty/3452,koleda_zloczowska',
-              'http://spiewnikreligijny.pl/teksty/4497,kolednicy_wedrownicy',
-              'http://spiewnikreligijny.pl/teksty/3291,kolysanie',
-              'http://spiewnikreligijny.pl/teksty/621,kolysanka_dla_tesi',
-              'http://spiewnikreligijny.pl/teksty/616,koniec_swiata',
-              'http://spiewnikreligijny.pl/teksty/3931,krew_chrystusa',
-              'http://spiewnikreligijny.pl/teksty/2899,krew_i_milosc',
-              'http://spiewnikreligijny.pl/teksty/3126,krok_krok',
-              'http://spiewnikreligijny.pl/teksty/351,kropla_morza',
-              'http://spiewnikreligijny.pl/teksty/3310,krol_krolow',
-              'http://spiewnikreligijny.pl/teksty/148,krola_wznosza_sie_ramiona',
-              'http://spiewnikreligijny.pl/teksty/4899,krolowa_aniolow',
-              'http://spiewnikreligijny.pl/teksty/747,krolowej_anielskiej_spiewajmy',
-              'http://spiewnikreligijny.pl/teksty/1184,krolowi_wiekow',
-              'http://spiewnikreligijny.pl/teksty/4123,krolowo_mych_snow',
-              'http://spiewnikreligijny.pl/teksty/4471,kroluj_nam_chryste',
-              'http://spiewnikreligijny.pl/teksty/153,krzyz',
-              'http://spiewnikreligijny.pl/teksty/152,krzyz_a_na_nim_bog',
-              'http://spiewnikreligijny.pl/teksty/3184,krzyz_jest_zrodlem',
-              'http://spiewnikreligijny.pl/teksty/147,krzyzu_chrystusa_badzze_pochwalony',
-              'http://spiewnikreligijny.pl/teksty/137,krzyzu_swiety_(hymn_o_krzyzu)',
-              'http://spiewnikreligijny.pl/teksty/146,krzyzu_swiety_nade_wszystko',
-              'http://spiewnikreligijny.pl/teksty/149,krzyzu_moj_krzyzu',
-              'http://spiewnikreligijny.pl/teksty/3259,kto_bogu_zaufa',
-              'http://spiewnikreligijny.pl/teksty/569,kto_nas_odlaczy',
-              'http://spiewnikreligijny.pl/teksty/2208,kto_odczuwa_pragnienie',
-              'http://spiewnikreligijny.pl/teksty/3272,kto_sadzi_ten_bladzi',
-              'http://spiewnikreligijny.pl/teksty/1888,ktos_cie_dzisiaj_wola',
-              'http://spiewnikreligijny.pl/teksty/1262,ktos_skarbow',
-              'http://spiewnikreligijny.pl/teksty/4864,ktory_skrzywdziles',
-              'http://spiewnikreligijny.pl/teksty/151,ktorys_za_nas_cierpial_rany',
-              'http://spiewnikreligijny.pl/teksty/4851,ktoryz_bog',
-              'http://spiewnikreligijny.pl/teksty/2177,ktoz_jak_bog',
-              'http://spiewnikreligijny.pl/teksty/350,ktoz_o_tej_dobie',
-              'http://spiewnikreligijny.pl/teksty/4557,ku_maryi',
-              'http://spiewnikreligijny.pl/teksty/748,ku_tobie_panie',
-              'http://spiewnikreligijny.pl/teksty/1920,kwiaty_komos', 'http://spiewnikreligijny.pl/teksty/3255,kyrie',
-              'http://spiewnikreligijny.pl/teksty/1222,la_sagrada_familia',
-              'http://spiewnikreligijny.pl/teksty/2964,lament_piotra',
-              'http://spiewnikreligijny.pl/teksty/156,lament_serdeczny',
-              'http://spiewnikreligijny.pl/teksty/4772,lamentacja',
-              'http://spiewnikreligijny.pl/teksty/2321,laudate_omnes_gentes',
-              'http://spiewnikreligijny.pl/teksty/751,laudato_sii',
-              'http://spiewnikreligijny.pl/teksty/2432,lec_koledo',
-              'http://spiewnikreligijny.pl/teksty/3436,lec_koledo',
-              'http://spiewnikreligijny.pl/teksty/2524,lec_koledo_po_gorach_i_dolinach_',
-              'http://spiewnikreligijny.pl/teksty/4148,lec_koledo_w_swiat',
-              'http://spiewnikreligijny.pl/teksty/3408,lew_hadasz', 'http://spiewnikreligijny.pl/teksty/4780,lew_judy',
-              'http://spiewnikreligijny.pl/teksty/752,licze_na_ciebie_ojcze',
-              'http://spiewnikreligijny.pl/teksty/4794,list', 'http://spiewnikreligijny.pl/teksty/1999,list_do_boga',
-              'http://spiewnikreligijny.pl/teksty/595,list_do_kosciola',
-              'http://spiewnikreligijny.pl/teksty/754,list_do_pana_boga', 'http://spiewnikreligijny.pl/teksty/755,lisc',
-              'http://spiewnikreligijny.pl/teksty/753,litania_loretanska',
-              'http://spiewnikreligijny.pl/teksty/3158,litania_pokutna',
-              'http://spiewnikreligijny.pl/teksty/4749,lokomotywa_z_ogloszenia_(perfect)',
-              'http://spiewnikreligijny.pl/teksty/936,lolek',
-              'http://spiewnikreligijny.pl/teksty/3120,lud_twoj_panie_lud_pielgrzymi',
-              'http://spiewnikreligijny.pl/teksty/1508,ludu_kaplanski',
-              'http://spiewnikreligijny.pl/teksty/155,ludu_moj_ludu', 'http://spiewnikreligijny.pl/teksty/756,ludzie',
-              'http://spiewnikreligijny.pl/teksty/757,ludzie_ludzie',
-              'http://spiewnikreligijny.pl/teksty/3563,lulaj_go_matko_lulaj',
-              'http://spiewnikreligijny.pl/teksty/400,lulajze_jezu_moj_drogi',
-              'http://spiewnikreligijny.pl/teksty/399,lulajze_jezuniu',
-              'http://spiewnikreligijny.pl/teksty/3536,lulilaj_spij_malenki',
-              'http://spiewnikreligijny.pl/teksty/750,laska_jego_trwa_na_wieki',
-              'http://spiewnikreligijny.pl/teksty/4058,laska_twoja_panie',
-              'http://spiewnikreligijny.pl/teksty/749,laskawosc_twoja_panie',
-              'http://spiewnikreligijny.pl/teksty/474,laska_jestesmy_zbawieni',
-              'http://spiewnikreligijny.pl/teksty/4819,lazenie_po_nocy',
-              'http://spiewnikreligijny.pl/teksty/1091,ma_dusza_pragnie_boga',
-              'http://spiewnikreligijny.pl/teksty/759,madonna', 'http://spiewnikreligijny.pl/teksty/914,magdalena',
-              'http://spiewnikreligijny.pl/teksty/760,magnificat', 'http://spiewnikreligijny.pl/teksty/762,magnificat',
-              'http://spiewnikreligijny.pl/teksty/2043,magnificat',
-              'http://spiewnikreligijny.pl/teksty/761,magnificat_(uwielbia_dusza_moja_pana)',
-              'http://spiewnikreligijny.pl/teksty/573,magnificat_(uwielbia_dusza_moja)',
-              'http://spiewnikreligijny.pl/teksty/2889,majowa_nasza_pani',
-              'http://spiewnikreligijny.pl/teksty/558,malarza_swiata',
-              'http://spiewnikreligijny.pl/teksty/412,malenka_milosc',
-              'http://spiewnikreligijny.pl/teksty/4376,mala__milosc',
-              'http://spiewnikreligijny.pl/teksty/537,male_milosci',
-              'http://spiewnikreligijny.pl/teksty/4626,maly_zbawiciel',
-              'http://spiewnikreligijny.pl/teksty/4347,mam_od_dzisiaj_przyjaciela',
-              'http://spiewnikreligijny.pl/teksty/763,mama', 'http://spiewnikreligijny.pl/teksty/1932,mama_krolowa',
-              'http://spiewnikreligijny.pl/teksty/4895,mamie', 'http://spiewnikreligijny.pl/teksty/1808,mamona',
-              'http://spiewnikreligijny.pl/teksty/73,mamy_czas', 'http://spiewnikreligijny.pl/teksty/3479,manifest',
-              'http://spiewnikreligijny.pl/teksty/566,marana_tha_(duch_pana_na_nim_spoczal)',
-              'http://spiewnikreligijny.pl/teksty/4202,marana_tha_(gdy_jak_pustynia)',
-              'http://spiewnikreligijny.pl/teksty/758,marana_tha_panie_nasz_przyjdz',
-              'http://spiewnikreligijny.pl/teksty/764,marana_tha_przyjdz_jezu_panie',
-              'http://spiewnikreligijny.pl/teksty/4773,maria_maria',
-              'http://spiewnikreligijny.pl/teksty/1984,maria_regina',
-              'http://spiewnikreligijny.pl/teksty/4861,marnosc_nad_marnosciami',
-              'http://spiewnikreligijny.pl/teksty/502,maryja',
-              'http://spiewnikreligijny.pl/teksty/4113,maryja_cudne_imie',
-              'http://spiewnikreligijny.pl/teksty/4325,maryjo_chce_zamieszkac_z_toba',
-              'http://spiewnikreligijny.pl/teksty/3728,maryjo_ja_twe_dziecie',
-              'http://spiewnikreligijny.pl/teksty/3351,maryjo_jesienna_i_rozancowa',
-              'http://spiewnikreligijny.pl/teksty/2969,maryjo_kocham_cie',
-              'http://spiewnikreligijny.pl/teksty/899,maryjo_mamo',
-              'http://spiewnikreligijny.pl/teksty/2041,maryjo_matko_mojego_wezwania',
-              'http://spiewnikreligijny.pl/teksty/162,maryjo_matko_tych_co_umieraja_razem_z_panem',
-              'http://spiewnikreligijny.pl/teksty/2606,maryjo_moje_zycie',
-              'http://spiewnikreligijny.pl/teksty/4106,maryjo_nasza_krolowa',
-              'http://spiewnikreligijny.pl/teksty/1189,maryjo_sliczna_pani',
-              'http://spiewnikreligijny.pl/teksty/4105,maryjo_tys_nasza_nadzieja',
-              'http://spiewnikreligijny.pl/teksty/3200,maryjo_wez_mnie_za_reke',
-              'http://spiewnikreligijny.pl/teksty/4456,maryjo_z_la_salette',
-              'http://spiewnikreligijny.pl/teksty/930,maryjo_maryjo',
-              'http://spiewnikreligijny.pl/teksty/413,marze_o_podrozach',
-              'http://spiewnikreligijny.pl/teksty/161,mater_dolorosa',
-              'http://spiewnikreligijny.pl/teksty/2656,matka_bolejaca',
-              'http://spiewnikreligijny.pl/teksty/2673,matka_ktora_pod_krzyzem_stala',
-              'http://spiewnikreligijny.pl/teksty/2790,matka_ktora_wszystko_rozumie',
-              'http://spiewnikreligijny.pl/teksty/1891,matko_boga_krolowo_swiata',
-              'http://spiewnikreligijny.pl/teksty/661,matko_ja_wielbie_cie',
-              'http://spiewnikreligijny.pl/teksty/1892,matko_ktora_nas_znasz',
-              'http://spiewnikreligijny.pl/teksty/4134,matko_ma',
-              'http://spiewnikreligijny.pl/teksty/4104,matko_ma_kosciele_moj',
-              'http://spiewnikreligijny.pl/teksty/4550,matko_milosci',
-              'http://spiewnikreligijny.pl/teksty/980,matko_najswietsza',
-              'http://spiewnikreligijny.pl/teksty/1605,matko_o_twarzy_czarnej',
-              'http://spiewnikreligijny.pl/teksty/606,matulu_moja',
-              'http://spiewnikreligijny.pl/teksty/38,madrosci_ktora_z_ust_bozych_wyplywasz',
-              'http://spiewnikreligijny.pl/teksty/1270,memu_bogu', 'http://spiewnikreligijny.pl/teksty/1890,menuet',
-              'http://spiewnikreligijny.pl/teksty/4795,mesjasz_jeszua',
-              'http://spiewnikreligijny.pl/teksty/411,medrcy_swiata_monarchowie',
-              'http://spiewnikreligijny.pl/teksty/540,mialbys_to_wszystko',
-              'http://spiewnikreligijny.pl/teksty/164,mialem_ci_ja_w_sercu',
-              'http://spiewnikreligijny.pl/teksty/3881,miej_nadzieje',
-              'http://spiewnikreligijny.pl/teksty/4884,milion_dolarow',
-              'http://spiewnikreligijny.pl/teksty/951,milosierdzie',
-              'http://spiewnikreligijny.pl/teksty/159,milosierdzie_boze',
-              'http://spiewnikreligijny.pl/teksty/2865,milosierdzie_twoje_panie',
-              'http://spiewnikreligijny.pl/teksty/163,milosierdziem_twoim_panie',
-              'http://spiewnikreligijny.pl/teksty/4059,milosierdziem_twoim_panie',
-              'http://spiewnikreligijny.pl/teksty/3410,milosc', 'http://spiewnikreligijny.pl/teksty/1619,milosc_boza',
-              'http://spiewnikreligijny.pl/teksty/972,milosc_da',
-              'http://spiewnikreligijny.pl/teksty/1261,milosc_do_ciebie_jest_moja_piesnia',
-              'http://spiewnikreligijny.pl/teksty/4544,milosc_dobra_jest',
-              'http://spiewnikreligijny.pl/teksty/607,milosc_dziki_wiatr',
-              'http://spiewnikreligijny.pl/teksty/3164,milosc_ii',
-              'http://spiewnikreligijny.pl/teksty/3390,milosc_jedyna',
-              'http://spiewnikreligijny.pl/teksty/3360,milosc_ktora_jest_bog',
-              'http://spiewnikreligijny.pl/teksty/3153,milosc_nieprzyjaciol',
-              'http://spiewnikreligijny.pl/teksty/2789,milosc_pana_niepojeta_jest',
-              'http://spiewnikreligijny.pl/teksty/1513,milosc_twa',
-              'http://spiewnikreligijny.pl/teksty/3913,milosc_twa_przyszla_na_swiat',
-              'http://spiewnikreligijny.pl/teksty/4305,milosc_twa_tak_cudowna_jest',
-              'http://spiewnikreligijny.pl/teksty/3147,milosc_ukrzyzowana',
-              'http://spiewnikreligijny.pl/teksty/551,milowaniem_serca_w_nas_obudzisz',
-              'http://spiewnikreligijny.pl/teksty/3113,milujcie_sie',
-              'http://spiewnikreligijny.pl/teksty/1315,mily_moj',
-              'http://spiewnikreligijny.pl/teksty/1722,minal_tydzien', 'http://spiewnikreligijny.pl/teksty/765,miriam',
-              'http://spiewnikreligijny.pl/teksty/766,miriam_loretanska',
-              'http://spiewnikreligijny.pl/teksty/767,misericordias_domini',
-              'http://spiewnikreligijny.pl/teksty/523,misja',
-              'http://spiewnikreligijny.pl/teksty/410,mizerna_cicha_stajenka_licha',
-              'http://spiewnikreligijny.pl/teksty/1893,moca_swa_obdarz_nas',
-              'http://spiewnikreligijny.pl/teksty/1807,modlitwa',
-              'http://spiewnikreligijny.pl/teksty/160,modlitwa_arcykaplanska',
-              'http://spiewnikreligijny.pl/teksty/1872,modlitwa_w_ciszy',
-              'http://spiewnikreligijny.pl/teksty/1841,modlitwa_za_rodziny_',
-              'http://spiewnikreligijny.pl/teksty/4479,modlitwa_zeglarza',
-              'http://spiewnikreligijny.pl/teksty/608,moga_sobie_mowic_to_co_chca',
-              'http://spiewnikreligijny.pl/teksty/1330,moge_dzielic_sie_wiatrem_i_chlebem',
-              'http://spiewnikreligijny.pl/teksty/939,moge_pojsc_dokad_tylko_chce__',
-              'http://spiewnikreligijny.pl/teksty/768,moi_przyjaciele',
-              'http://spiewnikreligijny.pl/teksty/4854,moja_gora',
-              'http://spiewnikreligijny.pl/teksty/3659,moja_modlitwa',
-              'http://spiewnikreligijny.pl/teksty/4593,moja_pokorna_prosba_do_matki_bozej',
-              'http://spiewnikreligijny.pl/teksty/2278,moje_marzenia',
-              'http://spiewnikreligijny.pl/teksty/1521,moje_miasto_wadowice',
-              'http://spiewnikreligijny.pl/teksty/85,moje_oczy_to_widzialy',
-              'http://spiewnikreligijny.pl/teksty/2796,moje_owce',
-              'http://spiewnikreligijny.pl/teksty/4193,moje_serce_nalezy_do_jezusa',
-              'http://spiewnikreligijny.pl/teksty/4585,moje_serce_sie_w_tobie_zakochalo',
-              'http://spiewnikreligijny.pl/teksty/90,moje_serce_spi',
-              'http://spiewnikreligijny.pl/teksty/1500,moje_wedrowanie',
-              'http://spiewnikreligijny.pl/teksty/3961,moje_wedrowanie',
-              'http://spiewnikreligijny.pl/teksty/4122,mosty', 'http://spiewnikreligijny.pl/teksty/1206,moze_daleko',
-              'http://spiewnikreligijny.pl/teksty/3829,mozesz_panie',
-              'http://spiewnikreligijny.pl/teksty/3920,moj_jezu_juz_sie_narodziles',
-              'http://spiewnikreligijny.pl/teksty/769,moj_jezu_moj_zbawco',
-              'http://spiewnikreligijny.pl/teksty/1894,moj_jezus_krolem_krolow_jest',
-              'http://spiewnikreligijny.pl/teksty/3761,moj_jezus_zbawiciel',
-              'http://spiewnikreligijny.pl/teksty/2404,moj_lad', 'http://spiewnikreligijny.pl/teksty/503,moj_mily',
-              'http://spiewnikreligijny.pl/teksty/158,moj_mistrzu',
-              'http://spiewnikreligijny.pl/teksty/3389,moj_pan_kocha_mnie',
-              'http://spiewnikreligijny.pl/teksty/770,moj_zbawiciel',
-              'http://spiewnikreligijny.pl/teksty/3013,mow_do_mego_serca',
-              'http://spiewnikreligijny.pl/teksty/2274,mow_do_mnie_panie',
-              'http://spiewnikreligijny.pl/teksty/3264,mowie_wam', 'http://spiewnikreligijny.pl/teksty/4005,mowisz',
-              'http://spiewnikreligijny.pl/teksty/2403,mowisz_ze_cisza_moim_domem',
-              'http://spiewnikreligijny.pl/teksty/1895,mrok_ogarnal',
-              'http://spiewnikreligijny.pl/teksty/997,mrozna_cisza_',
-              'http://spiewnikreligijny.pl/teksty/771,musimy_siac',
-              'http://spiewnikreligijny.pl/teksty/1244,my_izra_izra_izrael',
-              'http://spiewnikreligijny.pl/teksty/3954,my_jego_dzieci',
-              'http://spiewnikreligijny.pl/teksty/4514,my_tez_pastuszkowie_',
-              'http://spiewnikreligijny.pl/teksty/1806,my_ufamy_tobie',
-              'http://spiewnikreligijny.pl/teksty/52,na_adwent_naszego_czekania',
-              'http://spiewnikreligijny.pl/teksty/418,na_boze_narodzenie',
-              'http://spiewnikreligijny.pl/teksty/168,na_drodze_cie_jezu_spotykam',
-              'http://spiewnikreligijny.pl/teksty/82,na_drugi_brzeg',
-              'http://spiewnikreligijny.pl/teksty/1141,na_golgocie',
-              'http://spiewnikreligijny.pl/teksty/169,na_golgocie_stoi_krzyz',
-              'http://spiewnikreligijny.pl/teksty/1067,na_jeziorze_wielka_burza',
-              'http://spiewnikreligijny.pl/teksty/4216,na_kiju_gwiazda',
-              'http://spiewnikreligijny.pl/teksty/4509,na_kopie_siana_lezalem_z_rana',
-              'http://spiewnikreligijny.pl/teksty/3011,na_krawedzi_wiecznosci',
-              'http://spiewnikreligijny.pl/teksty/1537,na_krzyzu',
-              'http://spiewnikreligijny.pl/teksty/3189,na_nowo_stwarzaj_mnie',
-              'http://spiewnikreligijny.pl/teksty/4491,na_pasterke',
-              'http://spiewnikreligijny.pl/teksty/2683,na_przydroznym_krzyzu',
-              'http://spiewnikreligijny.pl/teksty/2676,na_rozdrozu_drog',
-              'http://spiewnikreligijny.pl/teksty/4871,na_skraju',
-              'http://spiewnikreligijny.pl/teksty/2772,na_skrzyzowaniu_swiata_drog',
-              'http://spiewnikreligijny.pl/teksty/2373,na_tym_niebie',
-              'http://spiewnikreligijny.pl/teksty/4283,na_via_dolorosa',
-              'http://spiewnikreligijny.pl/teksty/3307,na_wieki_bedzie_trwac_chwala_boza',
-              'http://spiewnikreligijny.pl/teksty/4680,na_zachod_dzien_sie_chyli',
-              'http://spiewnikreligijny.pl/teksty/3508,nad_betlejem_',
-              'http://spiewnikreligijny.pl/teksty/772,nad_jordanem',
-              'http://spiewnikreligijny.pl/teksty/4331,nad_katyniem_nad_smolenskiem',
-              'http://spiewnikreligijny.pl/teksty/773,nad_polska_chmury_rozprasza',
-              'http://spiewnikreligijny.pl/teksty/4789,nad_rzekami_babilonu',
-              'http://spiewnikreligijny.pl/teksty/3204,nad_rzekami_babilonu',
-              'http://spiewnikreligijny.pl/teksty/4629,nad_uspiona_galilea',
-              'http://spiewnikreligijny.pl/teksty/3387,nad_wodami_babilonu',
-              'http://spiewnikreligijny.pl/teksty/3981,nadszedl_juz_czas',
-              'http://spiewnikreligijny.pl/teksty/1059,nadzieja', 'http://spiewnikreligijny.pl/teksty/4862,nadzieja',
-              'http://spiewnikreligijny.pl/teksty/1146,nadzieja_(pater_hemon)',
-              'http://spiewnikreligijny.pl/teksty/165,najdrozszy_jezu',
-              'http://spiewnikreligijny.pl/teksty/61,najlepsza_modlitwa_na_swiecie',
-              'http://spiewnikreligijny.pl/teksty/4814,najlepsze_zyczenia',
-              'http://spiewnikreligijny.pl/teksty/4750,najlepszy_na_swiecie',
-              'http://spiewnikreligijny.pl/teksty/2843,najmilszy_panie_moj',
-              'http://spiewnikreligijny.pl/teksty/3533,najpierw_aniol_jej_zwiastowal',
-              'http://spiewnikreligijny.pl/teksty/4529,najswietsza_dziewico',
-              'http://spiewnikreligijny.pl/teksty/3078,najswietsze_serce',
-              'http://spiewnikreligijny.pl/teksty/4753,najwiekszy_gazda',
-              'http://spiewnikreligijny.pl/teksty/3734,nakarm_nas_panie',
-              'http://spiewnikreligijny.pl/teksty/944,nakarmiles_mnie',
-              'http://spiewnikreligijny.pl/teksty/504,napelnijcie_stagwie_woda',
-              'http://spiewnikreligijny.pl/teksty/414,narodzil_sie_jezus_chrystus',
-              'http://spiewnikreligijny.pl/teksty/417,narodzil_sie_jezus_w_stajni_ubogi',
-              'http://spiewnikreligijny.pl/teksty/415,narodzil_sie_nam_zbawiciel',
-              'http://spiewnikreligijny.pl/teksty/2379,narodzil_sie_w_stajence',
-              'http://spiewnikreligijny.pl/teksty/4722,nasz_bog_jest_wielki',
-              'http://spiewnikreligijny.pl/teksty/3093,nasz_pan',
-              'http://spiewnikreligijny.pl/teksty/3647,nasz_pan_jest_potezny',
-              'http://spiewnikreligijny.pl/teksty/3378,nasz_pasterz_odszedl',
-              'http://spiewnikreligijny.pl/teksty/3166,nasza_droga',
-              'http://spiewnikreligijny.pl/teksty/1254,nasze_plany',
-              'http://spiewnikreligijny.pl/teksty/3195,naucz_mnie_milosci_krzyza',
-              'http://spiewnikreligijny.pl/teksty/1896,naucz_mnie_panie_swojej_drogi',
-              'http://spiewnikreligijny.pl/teksty/3548,naucz_mnie_panie_wyslawiac_cie',
-              'http://spiewnikreligijny.pl/teksty/174,naucz_nas_panie',
-              'http://spiewnikreligijny.pl/teksty/2020,nic_nie_musisz_mowic',
-              'http://spiewnikreligijny.pl/teksty/3244,nic_pewnego',
-              'http://spiewnikreligijny.pl/teksty/524,nie_boj_sie',
-              'http://spiewnikreligijny.pl/teksty/2017,nie_boj_sie_bo_jestem_z_toba',
-              'http://spiewnikreligijny.pl/teksty/1719,nie_boj_sie_milosci_ktora_wola_cie',
-              'http://spiewnikreligijny.pl/teksty/895,nie_boj_sie_nie_lekaj_sie',
-              'http://spiewnikreligijny.pl/teksty/505,nie_boj_sie_wyplyn_na_glebie',
-              'http://spiewnikreligijny.pl/teksty/506,nie_bojcie_sie',
-              'http://spiewnikreligijny.pl/teksty/416,nie_bylo_miejsca_dla_ciebie',
-              'http://spiewnikreligijny.pl/teksty/1897,nie_chce_chodzic',
-              'http://spiewnikreligijny.pl/teksty/918,nie_chce_wecej',
-              'http://spiewnikreligijny.pl/teksty/2591,nie_chce_wiecej_by',
-              'http://spiewnikreligijny.pl/teksty/4735,nie_cykali_sie',
-              'http://spiewnikreligijny.pl/teksty/534,nie_dzieje_sie_nic',
-              'http://spiewnikreligijny.pl/teksty/4790,nie_jestem_sam_(odpowiedz)',
-              'http://spiewnikreligijny.pl/teksty/2361,nie_jestes_sam',
-              'http://spiewnikreligijny.pl/teksty/3274,nie_jestes_sam_nawet_jesli_boisz_sie',
-              'http://spiewnikreligijny.pl/teksty/2824,nie_karc_mnie_panie',
-              'http://spiewnikreligijny.pl/teksty/2914,nie_lekajcie_sie',
-              'http://spiewnikreligijny.pl/teksty/969,nie_lekaj_sie',
-              'http://spiewnikreligijny.pl/teksty/93,nie_lekaj_sie',
-              'http://spiewnikreligijny.pl/teksty/597,nie_lekaj_sie_(kiedy_wilk_zamieszka)',
-              'http://spiewnikreligijny.pl/teksty/2991,nie_lekaj_sie_moj_slugo',
-              'http://spiewnikreligijny.pl/teksty/3433,nie_lekajcie_sie',
-              'http://spiewnikreligijny.pl/teksty/2179,nie_lekajcie_sie_(hymn)',
-              'http://spiewnikreligijny.pl/teksty/4548,nie_lekajcie_sie_byc_swietymi',
-              'http://spiewnikreligijny.pl/teksty/2302,nie_lekajcie_sie_i',
-              'http://spiewnikreligijny.pl/teksty/3150,nie_ma_cie_panie',
-              'http://spiewnikreligijny.pl/teksty/545,nie_ma_mnie_bez_ciebie',
-              'http://spiewnikreligijny.pl/teksty/3271,nie_ma_nic_niemozliwego',
-              'http://spiewnikreligijny.pl/teksty/557,nie_ma_nieba_na_ziemi',
-              'http://spiewnikreligijny.pl/teksty/1702,nie_mam_nic_co_bym_mogl_tobie_dac',
-              'http://spiewnikreligijny.pl/teksty/2895,nie_mozna_zyc_bez_swiatla',
-              'http://spiewnikreligijny.pl/teksty/3314,nie_nam_lecz_twemu_imieniu',
-              'http://spiewnikreligijny.pl/teksty/1720,nie_odchodz_panie',
-              'http://spiewnikreligijny.pl/teksty/167,nie_opuszczaj_nas',
-              'http://spiewnikreligijny.pl/teksty/4857,nie_polegam_na_sobie',
-              'http://spiewnikreligijny.pl/teksty/4835,nie_powiem_nic',
-              'http://spiewnikreligijny.pl/teksty/779,nie_rzucim_chryste',
-              'http://spiewnikreligijny.pl/teksty/1038,nie_scigaj_sie_z_miloscia',
-              'http://spiewnikreligijny.pl/teksty/423,nie_tylko_tam_w_szopce',
-              'http://spiewnikreligijny.pl/teksty/780,nie_umiem_dziekowac_ci_panie',
-              'http://spiewnikreligijny.pl/teksty/3306,nie_umiem_dziekowac_ci_panie_ii',
-              'http://spiewnikreligijny.pl/teksty/3411,nie_umre',
-              'http://spiewnikreligijny.pl/teksty/781,nie_zabieraj_mi_strun',
-              'http://spiewnikreligijny.pl/teksty/968,nie_zabraknie_mi_nigdy_chleba',
-              'http://spiewnikreligijny.pl/teksty/4085,nie_zabraniajcie_im_przychodzic_do_mnie',
-              'http://spiewnikreligijny.pl/teksty/2866,nie_zamykajmy_serc',
-              'http://spiewnikreligijny.pl/teksty/166,nie_zdejme_krzyza_z_mojej_sciany',
-              'http://spiewnikreligijny.pl/teksty/5,nie_zna_smierci_pan_zywota',
-              'http://spiewnikreligijny.pl/teksty/4806,nie_nie_nie',
-              'http://spiewnikreligijny.pl/teksty/173,nie_nie_nie_umieraj',
-              'http://spiewnikreligijny.pl/teksty/609,niebianskie_nawolywanie',
-              'http://spiewnikreligijny.pl/teksty/39,niebiosa_rose_spuscie_nam_z_gory',
-              'http://spiewnikreligijny.pl/teksty/3015,niebo_dla_mnie',
-              'http://spiewnikreligijny.pl/teksty/904,niebo_juz_pelne_gwiazd',
-              'http://spiewnikreligijny.pl/teksty/172,niebo_najwspanialsze_z_miejsc',
-              'http://spiewnikreligijny.pl/teksty/78,niebo_niebieskie',
-              'http://spiewnikreligijny.pl/teksty/777,niech_bedzie_chwala',
-              'http://spiewnikreligijny.pl/teksty/778,niech_cie_pan_blogoslawi',
-              'http://spiewnikreligijny.pl/teksty/1898,niech_milosc_twoja',
-              'http://spiewnikreligijny.pl/teksty/1221,niech_moje_milczenie_strzeze_ofiary',
-              'http://spiewnikreligijny.pl/teksty/1966,niech_nas_ogarnie_',
-              'http://spiewnikreligijny.pl/teksty/1271,niech_nas_ogarnie_laska_panie_twa',
-              'http://spiewnikreligijny.pl/teksty/4821,niech_nie_kusi',
-              'http://spiewnikreligijny.pl/teksty/2870,niech_oblicze_twe',
-              'http://spiewnikreligijny.pl/teksty/1846,niech_pan_udzieli_mocy_swemu_ludowi',
-              'http://spiewnikreligijny.pl/teksty/3279,niech_sie_dzieje',
-              'http://spiewnikreligijny.pl/teksty/3300,niech_sie_stanie',
-              'http://spiewnikreligijny.pl/teksty/2940,niech_sie_stanie_twoja_wola',
-              'http://spiewnikreligijny.pl/teksty/2267,niech_strumienie_zywej_wody',
-              'http://spiewnikreligijny.pl/teksty/4850,niech_twa_chwala',
-              'http://spiewnikreligijny.pl/teksty/2807,niech_twoj_swiety_duch',
-              'http://spiewnikreligijny.pl/teksty/13,niech_w_swieto_radosne_(sekwencja)',
-              'http://spiewnikreligijny.pl/teksty/639,niech_zabrzmi_panu_chwala_w_niebiosach',
-              'http://spiewnikreligijny.pl/teksty/171,niech_ziemia_i_niebo',
-              'http://spiewnikreligijny.pl/teksty/419,niechaj_bedzie_glosno_wszedzie',
-              'http://spiewnikreligijny.pl/teksty/3054,niechaj_bedzie_na_wieki',
-              'http://spiewnikreligijny.pl/teksty/3366,niechaj_brzmi_tobie_chwaly_piesn',
-              'http://spiewnikreligijny.pl/teksty/867,niechaj_milosc_twa',
-              'http://spiewnikreligijny.pl/teksty/937,niechaj_plomien_',
-              'http://spiewnikreligijny.pl/teksty/554,niechaj_wielbi_ciebie_cala_ziemia',
-              'http://spiewnikreligijny.pl/teksty/775,niechaj_z_nami_bedzie_pan',
-              'http://spiewnikreligijny.pl/teksty/776,niechaj_zstapi_duch_twoj',
-              'http://spiewnikreligijny.pl/teksty/88,niedola_egipska_niewola',
-              'http://spiewnikreligijny.pl/teksty/97,niedziella',
-              'http://spiewnikreligijny.pl/teksty/614,niedzielny_przelot',
-              'http://spiewnikreligijny.pl/teksty/2737,niegodny', 'http://spiewnikreligijny.pl/teksty/4732,nieidealna',
-              'http://spiewnikreligijny.pl/teksty/420,niepojete_dary',
-              'http://spiewnikreligijny.pl/teksty/3309,niewyczerpane_sa_laski_twe_panie',
-              'http://spiewnikreligijny.pl/teksty/4744,niezwyciezony_(armia_)',
-              'http://spiewnikreligijny.pl/teksty/87,niezwykla_noc',
-              'http://spiewnikreligijny.pl/teksty/62,nigdy_nie_zostawisz_mnie',
-              'http://spiewnikreligijny.pl/teksty/2878,nigdy_z_krolami',
-              'http://spiewnikreligijny.pl/teksty/525,nikt_nie_jest_wyspa',
-              'http://spiewnikreligijny.pl/teksty/4767,nikt_nie_moze_sluzyc_dwom_panom',
-              'http://spiewnikreligijny.pl/teksty/2989,nim_swit_jasny_wstanie',
-              'http://spiewnikreligijny.pl/teksty/422,noc_ciemna_i_zimna',
-              'http://spiewnikreligijny.pl/teksty/170,noca_ogrod_oliwny',
-              'http://spiewnikreligijny.pl/teksty/4796,nowa_piesn',
-              'http://spiewnikreligijny.pl/teksty/4816,nowa_ziemia',
-              'http://spiewnikreligijny.pl/teksty/86,nowe_przykazanie',
-              'http://spiewnikreligijny.pl/teksty/421,nowy_rok_biezy',
-              'http://spiewnikreligijny.pl/teksty/424,nuzmy_wszyscy_zaspiewajmy',
-              'http://spiewnikreligijny.pl/teksty/290,o_boze_o_boze_panie_moj',
-              'http://spiewnikreligijny.pl/teksty/178,o_chryste_zwroc_twa_bolesna_glowe',
-              'http://spiewnikreligijny.pl/teksty/1176,o_chrystusowa_meko',
-              'http://spiewnikreligijny.pl/teksty/1866,o_cuda_cuda',
-              'http://spiewnikreligijny.pl/teksty/3299,o_czym_inni_tylko_snia',
-              'http://spiewnikreligijny.pl/teksty/21,o_dniu_radosny_pelen_chwal',
-              'http://spiewnikreligijny.pl/teksty/175,o_duszo_wszelka_nabozna',
-              'http://spiewnikreligijny.pl/teksty/3376,o_glowo_cos_zraniona',
-              'http://spiewnikreligijny.pl/teksty/176,o_glowo_uwienczona',
-              'http://spiewnikreligijny.pl/teksty/49,o_gospodzie_uwielbiona',
-              'http://spiewnikreligijny.pl/teksty/180,o_grobie_chwalebny',
-              'http://spiewnikreligijny.pl/teksty/425,o_gwiazdo_betlejemska',
-              'http://spiewnikreligijny.pl/teksty/3156,o_hostio_swieta',
-              'http://spiewnikreligijny.pl/teksty/3174,o_jahwe_boze_ojcow_moich',
-              'http://spiewnikreligijny.pl/teksty/2574,o_jezu',
-              'http://spiewnikreligijny.pl/teksty/3145,o_jezu_boze_utajony',
-              'http://spiewnikreligijny.pl/teksty/181,o_jezu_moj_zbawco_moj',
-              'http://spiewnikreligijny.pl/teksty/3019,o_jezu_w_hostii_bialej',
-              'http://spiewnikreligijny.pl/teksty/1899,o_jezu_cichy_i_pokorny',
-              'http://spiewnikreligijny.pl/teksty/426,o_jezu_nasz_zbawicielu',
-              'http://spiewnikreligijny.pl/teksty/429,o_jozefie_cos_szukal_gospody',
-              'http://spiewnikreligijny.pl/teksty/1036,o_krolowo',
-              'http://spiewnikreligijny.pl/teksty/971,o_krolu_chwaly',
-              'http://spiewnikreligijny.pl/teksty/179,o_krwi_najdrozsza',
-              'http://spiewnikreligijny.pl/teksty/428,o_malenki_jezu_moj',
-              'http://spiewnikreligijny.pl/teksty/3350,o_maryjo_krolowo_nasza',
-              'http://spiewnikreligijny.pl/teksty/785,o_maryjo_piekna_pani',
-              'http://spiewnikreligijny.pl/teksty/430,o_mili_krolowie',
-              'http://spiewnikreligijny.pl/teksty/184,o_moj_jezu',
-              'http://spiewnikreligijny.pl/teksty/3182,o_moj_najslodszy_jezu',
-              'http://spiewnikreligijny.pl/teksty/183,o_najswietsza_twarzy_mego_pana',
-              'http://spiewnikreligijny.pl/teksty/23,o_nasz_jezu_najlaskawszy',
-              'http://spiewnikreligijny.pl/teksty/786,o_pani_ufnosc_nasza',
-              'http://spiewnikreligijny.pl/teksty/1318,o_salutaris_hostia',
-              'http://spiewnikreligijny.pl/teksty/3232,o_stworzycielu_duchu_przyjdz',
-              'http://spiewnikreligijny.pl/teksty/1973,o_swieta_maryjo',
-              'http://spiewnikreligijny.pl/teksty/3364,o_tak_tak_tak_panie_mowie_tak',
-              'http://spiewnikreligijny.pl/teksty/864,o_wychwalajcie_go_wszystkie_narody',
-              'http://spiewnikreligijny.pl/teksty/790,o_wychwalajcie_go_wszystkie_narody',
-              'http://spiewnikreligijny.pl/teksty/4807,obietnica',
-              'http://spiewnikreligijny.pl/teksty/1034,oblubieniec',
-              'http://spiewnikreligijny.pl/teksty/3327,obudz_mnie',
-              'http://spiewnikreligijny.pl/teksty/3220,ocien_mnie_panie',
-              'http://spiewnikreligijny.pl/teksty/2717,ocierasz_moje_lzy',
-              'http://spiewnikreligijny.pl/teksty/3197,oczarowales_me_serce',
-              'http://spiewnikreligijny.pl/teksty/3811,oczekuje_wypatruje',
-              'http://spiewnikreligijny.pl/teksty/986,oczyszczenie',
-              'http://spiewnikreligijny.pl/teksty/3024,od_boga_jestem',
-              'http://spiewnikreligijny.pl/teksty/4882,od_poczatku_az_do_konca',
-              'http://spiewnikreligijny.pl/teksty/3097,oddaje_ci_cala_chwale_i_czesc',
-              'http://spiewnikreligijny.pl/teksty/1092,oddaje_ci_zycie_swe',
-              'http://spiewnikreligijny.pl/teksty/4856,oddal', 'http://spiewnikreligijny.pl/teksty/1015,odnajduje_cie',
-              'http://spiewnikreligijny.pl/teksty/2827,odrzuc_co_zle_czyn_dobro',
-              'http://spiewnikreligijny.pl/teksty/182,odszedl_pasterz_od_nas',
-              'http://spiewnikreligijny.pl/teksty/185,ofiara_krzyza_spelniona',
-              'http://spiewnikreligijny.pl/teksty/1859,ofiarowanie',
-              'http://spiewnikreligijny.pl/teksty/917,ofiaruje_tobie_panie',
-              'http://spiewnikreligijny.pl/teksty/782,ofiaruje_ci',
-              'http://spiewnikreligijny.pl/teksty/1667,ofiaruje_ci_panie',
-              'http://spiewnikreligijny.pl/teksty/2369,ofiaruje_panie_ci',
-              'http://spiewnikreligijny.pl/teksty/24,ofiarujmy_chwale_w_wierze',
-              'http://spiewnikreligijny.pl/teksty/4817,ogien', 'http://spiewnikreligijny.pl/teksty/2662,ogien_pal',
-              'http://spiewnikreligijny.pl/teksty/3379,oglaszamy_krolestwo_boze_w_nas',
-              'http://spiewnikreligijny.pl/teksty/1900,ogrody',
-              'http://spiewnikreligijny.pl/teksty/186,ogrodzie_oliwny', 'http://spiewnikreligijny.pl/teksty/1802,ogrod',
-              'http://spiewnikreligijny.pl/teksty/427,oj_maluski_maluski',
-              'http://spiewnikreligijny.pl/teksty/1023,ojciec_radosny',
-              'http://spiewnikreligijny.pl/teksty/3316,ojcze_chwala_tobie',
-              'http://spiewnikreligijny.pl/teksty/3140,ojcze_daj_mi_ducha',
-              'http://spiewnikreligijny.pl/teksty/1950,ojcze_nadziei',
-              'http://spiewnikreligijny.pl/teksty/187,ojcze_nasz',
-              'http://spiewnikreligijny.pl/teksty/3262,ojcze_nasz_ojcze_nasz',
-              'http://spiewnikreligijny.pl/teksty/783,ojcze_ty_kochasz_mnie',
-              'http://spiewnikreligijny.pl/teksty/191,ojcze_z_niebios',
-              'http://spiewnikreligijny.pl/teksty/190,ojcze_boze_wszechmogacy',
-              'http://spiewnikreligijny.pl/teksty/784,ojczyzno_ma',
-              'http://spiewnikreligijny.pl/teksty/984,okaz_mi_boze_milosierdzie',
-              'http://spiewnikreligijny.pl/teksty/3238,okulary_wiary',
-              'http://spiewnikreligijny.pl/teksty/76,olej_do_glowy', 'http://spiewnikreligijny.pl/teksty/1954,on',
-              'http://spiewnikreligijny.pl/teksty/1996,on_cie_wola', 'http://spiewnikreligijny.pl/teksty/1901,on_i_ja',
-              'http://spiewnikreligijny.pl/teksty/905,on_jest_moim_zyciem',
-              'http://spiewnikreligijny.pl/teksty/64,ona_powiedziala_tak',
-              'http://spiewnikreligijny.pl/teksty/3138,ona_stoi_w_slonce_obleczona',
-              'http://spiewnikreligijny.pl/teksty/1070,orzel',
-              'http://spiewnikreligijny.pl/teksty/1503,osiem_blogoslawienstw',
-              'http://spiewnikreligijny.pl/teksty/70,osiem_blogoslawienstw',
-              'http://spiewnikreligijny.pl/teksty/787,ostatnia_prosba',
-              'http://spiewnikreligijny.pl/teksty/189,oto_drzewo_krzyza',
-              'http://spiewnikreligijny.pl/teksty/1123,oto_ide',
-              'http://spiewnikreligijny.pl/teksty/2416,oto_idzie_moj_bog',
-              'http://spiewnikreligijny.pl/teksty/3329,oto_ja',
-              'http://spiewnikreligijny.pl/teksty/3039,oto_ja_poslij_mnie',
-              'http://spiewnikreligijny.pl/teksty/1002,oto_jest_dzien',
-              'http://spiewnikreligijny.pl/teksty/188,oto_jezus_umiera',
-              'http://spiewnikreligijny.pl/teksty/3057,oto_mesjasz',
-              'http://spiewnikreligijny.pl/teksty/192,oto_nadchodza_dni',
-              'http://spiewnikreligijny.pl/teksty/40,oto_pan_bog_przyjdzie',
-              'http://spiewnikreligijny.pl/teksty/788,oto_sa_baranki_mlode',
-              'http://spiewnikreligijny.pl/teksty/789,oto_stoje_u_drzwi_i_kolacze',
-              'http://spiewnikreligijny.pl/teksty/6,otrzyjcie_juz_lzy_placzacy',
-              'http://spiewnikreligijny.pl/teksty/2365,otwieram_dla_ciebie_serce_swe',
-              'http://spiewnikreligijny.pl/teksty/955,otworz_me_oczy',
-              'http://spiewnikreligijny.pl/teksty/1266,otworz_me_oczy_o_panie',
-              'http://spiewnikreligijny.pl/teksty/1798,otworz_oczy',
-              'http://spiewnikreligijny.pl/teksty/4870,otworz_swe_serce',
-              'http://spiewnikreligijny.pl/teksty/1011,otworzcie_bramy',
-              'http://spiewnikreligijny.pl/teksty/2975,otworzcie_drzwi',
-              'http://spiewnikreligijny.pl/teksty/3056,otworzcie_drzwi_chrystusowi',
-              'http://spiewnikreligijny.pl/teksty/2714,otworzcie_serca',
-              'http://spiewnikreligijny.pl/teksty/1267,otworzmy_serca', 'http://spiewnikreligijny.pl/teksty/1502,owca',
-              'http://spiewnikreligijny.pl/teksty/3753,owoce', 'http://spiewnikreligijny.pl/teksty/3377,ozdobo_nieba',
-              'http://spiewnikreligijny.pl/teksty/3124,pagorki', 'http://spiewnikreligijny.pl/teksty/100,palac',
-              'http://spiewnikreligijny.pl/teksty/194,pamietaj_czlowiecze_na_jezusa',
-              'http://spiewnikreligijny.pl/teksty/2706,pan_blisko_jest',
-              'http://spiewnikreligijny.pl/teksty/2983,pan_blisko_jest',
-              'http://spiewnikreligijny.pl/teksty/526,pan_jest_dobry',
-              'http://spiewnikreligijny.pl/teksty/1987,pan_jest_moca_swojego_ludu',
-              'http://spiewnikreligijny.pl/teksty/4859,pan_jest_moja_swiatloscia',
-              'http://spiewnikreligijny.pl/teksty/1989,pan_jest_pasterzem_moim',
-              'http://spiewnikreligijny.pl/teksty/940,pan_jest_pasterzem_moim_ii',
-              'http://spiewnikreligijny.pl/teksty/3381,pan_jest_wsrod_nas',
-              'http://spiewnikreligijny.pl/teksty/196,pan_jezus_grzechy_nasze',
-              'http://spiewnikreligijny.pl/teksty/1071,pan_kiedys_stanal_nad_brzegiem',
-              'http://spiewnikreligijny.pl/teksty/1988,pan_mnie_strzeze',
-              'http://spiewnikreligijny.pl/teksty/4015,pan_moj_wodz',
-              'http://spiewnikreligijny.pl/teksty/1625,pan_wywyzszony',
-              'http://spiewnikreligijny.pl/teksty/431,pan_z_nieba_i_z_lona_ojca_przychodzi',
-              'http://spiewnikreligijny.pl/teksty/3225,pan_zmartwychwstal',
-              'http://spiewnikreligijny.pl/teksty/14,pan_zmartwychwstal_i_jest_z_nami',
-              'http://spiewnikreligijny.pl/teksty/1247,panie_boze_do_ciebie_wolam',
-              'http://spiewnikreligijny.pl/teksty/1753,panie_boze_spraw',
-              'http://spiewnikreligijny.pl/teksty/200,panie_co_mieszkasz_na_wysokim_niebie',
-              'http://spiewnikreligijny.pl/teksty/3218,panie_daj_mi_laske_wiernej_modlitwy',
-              'http://spiewnikreligijny.pl/teksty/3149,panie_daj_mi_twej_wody',
-              'http://spiewnikreligijny.pl/teksty/2317,panie_do_kogoz_pojdziemy',
-              'http://spiewnikreligijny.pl/teksty/2948,panie_dobry_jak_chleb',
-              'http://spiewnikreligijny.pl/teksty/1709,panie_dzieki_ci',
-              'http://spiewnikreligijny.pl/teksty/1812,panie_jezu_moj',
-              'http://spiewnikreligijny.pl/teksty/1857,panie_jezu_przyjm',
-              'http://spiewnikreligijny.pl/teksty/3043,panie_jezu_zabierzemy_cie_do_domu',
-              'http://spiewnikreligijny.pl/teksty/3088,panie_ktory_czynisz_(akt_pokutny)',
-              'http://spiewnikreligijny.pl/teksty/3203,panie_moje_serce_male_jest',
-              'http://spiewnikreligijny.pl/teksty/1192,panie_moj',
-              'http://spiewnikreligijny.pl/teksty/4571,panie_moj_boze_wysluchaj_mnie',
-              'http://spiewnikreligijny.pl/teksty/3037,panie_my_twoje_dzieci',
-              'http://spiewnikreligijny.pl/teksty/2665,panie_nie_umiem_krzyza_niesc',
-              'http://spiewnikreligijny.pl/teksty/199,panie_nogi_twe_caluje',
-              'http://spiewnikreligijny.pl/teksty/3059,panie_pozostan',
-              'http://spiewnikreligijny.pl/teksty/4847,panie_pragne_ci_dziekowac',
-              'http://spiewnikreligijny.pl/teksty/2885,panie_pragnienia_ludzkich_serc',
-              'http://spiewnikreligijny.pl/teksty/1693,panie_prosze_przyjdz',
-              'http://spiewnikreligijny.pl/teksty/636,panie_przebacz_nam',
-              'http://spiewnikreligijny.pl/teksty/3157,panie_przepasz_mnie',
-              'http://spiewnikreligijny.pl/teksty/3210,panie_spraw_by_me_serce',
-              'http://spiewnikreligijny.pl/teksty/1530,panie_ty_nas_prowadz',
-              'http://spiewnikreligijny.pl/teksty/3207,panie_ty_znasz_moje_imie',
-              'http://spiewnikreligijny.pl/teksty/2941,panie_wielbie_cie',
-              'http://spiewnikreligijny.pl/teksty/1991,panie_wielbie_cie',
-              'http://spiewnikreligijny.pl/teksty/3221,panie_zmiluj_sie_nad_nami',
-              'http://spiewnikreligijny.pl/teksty/791,panie_dobry_jak_chleb',
-              'http://spiewnikreligijny.pl/teksty/932,panie_prosze_spraw',
-              'http://spiewnikreligijny.pl/teksty/198,panie_spojrz_stoje_tu',
-              'http://spiewnikreligijny.pl/teksty/1990,panie_twoj_tron',
-              'http://spiewnikreligijny.pl/teksty/195,panie_ty_widzisz_krzyza_sie_nie_lekam',
-              'http://spiewnikreligijny.pl/teksty/197,panie_ufam_milosierdziu_twemu',
-              'http://spiewnikreligijny.pl/teksty/2402,panu_chwala_i_czesc',
-              'http://spiewnikreligijny.pl/teksty/2954,panu_memu_oddac_chce',
-              'http://spiewnikreligijny.pl/teksty/906,panu_naszemu_piesni_grajcie',
-              'http://spiewnikreligijny.pl/teksty/3152,paschalny_deszcz',
-              'http://spiewnikreligijny.pl/teksty/1191,pasja',
-              'http://spiewnikreligijny.pl/teksty/433,pasterze_biezeli_gdy_glos_uslyszeli',
-              'http://spiewnikreligijny.pl/teksty/432,pasterze_mili_coscie_widzieli',
-              'http://spiewnikreligijny.pl/teksty/434,pasterzu_pasterzu_czy_widzisz',
-              'http://spiewnikreligijny.pl/teksty/3524,pastoralka_czudecka_',
-              'http://spiewnikreligijny.pl/teksty/435,pastuszek',
-              'http://spiewnikreligijny.pl/teksty/3083,patrz_swiety_bog_i_pan',
-              'http://spiewnikreligijny.pl/teksty/3482,patrzcie_bracia',
-              'http://spiewnikreligijny.pl/teksty/792,patrze_na_ptaki_i_lilie_polne',
-              'http://spiewnikreligijny.pl/teksty/2068,patrzysz_na_mnie',
-              'http://spiewnikreligijny.pl/teksty/2913,pentecoste',
-              'http://spiewnikreligijny.pl/teksty/4788,perelka_(psalm_71)',
-              'http://spiewnikreligijny.pl/teksty/4842,perla_oceanu',
-              'http://spiewnikreligijny.pl/teksty/1255,pewnej_nocy',
-              'http://spiewnikreligijny.pl/teksty/438,pierwsza_gwiazda',
-              'http://spiewnikreligijny.pl/teksty/2951,pierwsza_komunia_swieta',
-              'http://spiewnikreligijny.pl/teksty/4733,pies_i_kot',
-              'http://spiewnikreligijny.pl/teksty/3046,piesn_dziekczynienia',
-              'http://spiewnikreligijny.pl/teksty/1027,piesn_harcerzy_polskich',
-              'http://spiewnikreligijny.pl/teksty/995,piesn_jednosci',
-              'http://spiewnikreligijny.pl/teksty/3206,piesn_malej_duszy',
-              'http://spiewnikreligijny.pl/teksty/1385,piesn_na_uwielbienie_',
-              'http://spiewnikreligijny.pl/teksty/507,piesn_nad_piesniami',
-              'http://spiewnikreligijny.pl/teksty/508,piesn_nad_piesniami_2001',
-              'http://spiewnikreligijny.pl/teksty/1246,piesn_o_izraelu',
-              'http://spiewnikreligijny.pl/teksty/509,piesn_oblubienca',
-              'http://spiewnikreligijny.pl/teksty/1423,piesn_patnikow_o_matce_bozej_lichenskiej',
-              'http://spiewnikreligijny.pl/teksty/4808,piesn_pielgrzymow_(ile_mil)',
-              'http://spiewnikreligijny.pl/teksty/4785,piesn_rodzaju',
-              'http://spiewnikreligijny.pl/teksty/4119,piekna_tys_jest_jak_slonce',
-              'http://spiewnikreligijny.pl/teksty/1383,pij_ten_kielich',
-              'http://spiewnikreligijny.pl/teksty/205,pioseneczka',
-              'http://spiewnikreligijny.pl/teksty/4822,piosenka_dla_pozornie_opuszczonych',
-              'http://spiewnikreligijny.pl/teksty/3438,piosenka_na_powitanie_swietego_mikolaja',
-              'http://spiewnikreligijny.pl/teksty/4809,piosenka_nowa',
-              'http://spiewnikreligijny.pl/teksty/3289,piosenka_o_dobrym_pasterzu',
-              'http://spiewnikreligijny.pl/teksty/4797,piosenka_o_milosci',
-              'http://spiewnikreligijny.pl/teksty/1099,piosenka_pasterzy',
-              'http://spiewnikreligijny.pl/teksty/4798,piosenka_wigilijna',
-              'http://spiewnikreligijny.pl/teksty/1162,plurimos_annos',
-              'http://spiewnikreligijny.pl/teksty/201,placz_placz_kto_zyw',
-              'http://spiewnikreligijny.pl/teksty/1236,placzcie_anieli',
-              'http://spiewnikreligijny.pl/teksty/202,placzcie_anieli',
-              'http://spiewnikreligijny.pl/teksty/204,placzze_dzisiaj_duszo_wszelka',
-              'http://spiewnikreligijny.pl/teksty/3230,plomien_milosci',
-              'http://spiewnikreligijny.pl/teksty/436,plonie_gwiazda_nad_betlejem',
-              'http://spiewnikreligijny.pl/teksty/794,po_calej_ziemi_brzmi',
-              'http://spiewnikreligijny.pl/teksty/649,po_gorach_dolinach',
-              'http://spiewnikreligijny.pl/teksty/1098,po_koledzie',
-              'http://spiewnikreligijny.pl/teksty/793,po_krance_ziemi',
-              'http://spiewnikreligijny.pl/teksty/4771,po_tamtej_stronie_rzeki',
-              'http://spiewnikreligijny.pl/teksty/41,po_upadku_czlowieka_grzesznego',
-              'http://spiewnikreligijny.pl/teksty/3026,poblogoslaw_jezu_drogi',
-              'http://spiewnikreligijny.pl/teksty/1122,poblogoslaw_panie',
-              'http://spiewnikreligijny.pl/teksty/1179,poblogoslaw_panie_z_wysokiego_nieba',
-              'http://spiewnikreligijny.pl/teksty/1277,pochwalony__duch__jest__swiety',
-              'http://spiewnikreligijny.pl/teksty/1799,poczatek_dnia',
-              'http://spiewnikreligijny.pl/teksty/3079,pod_twa_obrone',
-              'http://spiewnikreligijny.pl/teksty/4845,podaj_dlon',
-              'http://spiewnikreligijny.pl/teksty/3127,podaj_reke_swemu_bratu',
-              'http://spiewnikreligijny.pl/teksty/2684,podaruj_podaruj_mi_panie',
-              'http://spiewnikreligijny.pl/teksty/510,podnies_mnie_jezu',
-              'http://spiewnikreligijny.pl/teksty/2962,pogoda_na_milosc',
-              'http://spiewnikreligijny.pl/teksty/98,pokochalem_cie',
-              'http://spiewnikreligijny.pl/teksty/1701,pokolenie_jp_ii',
-              'http://spiewnikreligijny.pl/teksty/206,pokora',
-              'http://spiewnikreligijny.pl/teksty/1197,pokorna_sluzebnico_pana',
-              'http://spiewnikreligijny.pl/teksty/553,pokoj_ludziom_dobrej_woli',
-              'http://spiewnikreligijny.pl/teksty/559,pokoj_moj_zostawiam',
-              'http://spiewnikreligijny.pl/teksty/3486,pokoj_wam',
-              'http://spiewnikreligijny.pl/teksty/795,polska_madonna',
-              'http://spiewnikreligijny.pl/teksty/3080,polskie_kwiaty',
-              'http://spiewnikreligijny.pl/teksty/2318,pomodl_sie_miriam',
-              'http://spiewnikreligijny.pl/teksty/2322,popatrz_jak_predko_mija_czas',
-              'http://spiewnikreligijny.pl/teksty/610,poprzez_zolty_kurz',
-              'http://spiewnikreligijny.pl/teksty/3453,poslal_go_do_nas_ojciec_bog',
-              'http://spiewnikreligijny.pl/teksty/2900,posluchaj_corko',
-              'http://spiewnikreligijny.pl/teksty/3063,posluchaj_mnie',
-              'http://spiewnikreligijny.pl/teksty/796,posylam_was',
-              'http://spiewnikreligijny.pl/teksty/193,posypmy_glowy_popiolem',
-              'http://spiewnikreligijny.pl/teksty/3242,poszukiwacze_skarbow',
-              'http://spiewnikreligijny.pl/teksty/797,poszukiwanie',
-              'http://spiewnikreligijny.pl/teksty/1654,poslij_nas',
-              'http://spiewnikreligijny.pl/teksty/439,pospieszcie_pastuszki_z_piosneczki',
-              'http://spiewnikreligijny.pl/teksty/3183,posrod_nocy',
-              'http://spiewnikreligijny.pl/teksty/965,potrac_te_cisze_',
-              'http://spiewnikreligijny.pl/teksty/1106,powiedz_ludziom_ze_kocham_ich',
-              'http://spiewnikreligijny.pl/teksty/207,powiedz_nam_krzyzu',
-              'http://spiewnikreligijny.pl/teksty/1496,powiedz_panie_dokad_zdazasz',
-              'http://spiewnikreligijny.pl/teksty/611,powiedz_powiedz_jak_to_jest',
-              'http://spiewnikreligijny.pl/teksty/437,powiedzcie_pasterze_mili',
-              'http://spiewnikreligijny.pl/teksty/3219,powierz_panu_swoja_droge',
-              'http://spiewnikreligijny.pl/teksty/1172,powierz_sie_matce',
-              'http://spiewnikreligijny.pl/teksty/1941,powietrzem_moim_jest',
-              'http://spiewnikreligijny.pl/teksty/4747,powinnosc_kurdupelka',
-              'http://spiewnikreligijny.pl/teksty/2837,powolales_mnie_do_zycia',
-              'http://spiewnikreligijny.pl/teksty/511,powstan_przyjaciolko_ma',
-              'http://spiewnikreligijny.pl/teksty/2047,pozdrawiam_ciebie_matko',
-              'http://spiewnikreligijny.pl/teksty/3172,pozdrawiam_cie_matko',
-              'http://spiewnikreligijny.pl/teksty/51,pozdrowiona_badz',
-              'http://spiewnikreligijny.pl/teksty/3018,pozostawiles_tu',
-              'http://spiewnikreligijny.pl/teksty/563,pozwol_bym_byl',
-              'http://spiewnikreligijny.pl/teksty/2324,pozwol_mi_przyjsc_do_ciebie',
-              'http://spiewnikreligijny.pl/teksty/203,pozwol_mi_twe_meki_spiewac',
-              'http://spiewnikreligijny.pl/teksty/598,pozwolcie_dzieciom',
-              'http://spiewnikreligijny.pl/teksty/1501,pojde_do_nieba_piechota',
-              'http://spiewnikreligijny.pl/teksty/3250,pojde_w_szeroki_swiat',
-              'http://spiewnikreligijny.pl/teksty/308,pojdziemy_bracia',
-              'http://spiewnikreligijny.pl/teksty/2679,pojdz_do_jezusa',
-              'http://spiewnikreligijny.pl/teksty/440,pojdzmy_wszyscy_do_stajenki',
-              'http://spiewnikreligijny.pl/teksty/441,pojdzmy_pojdzmy_az_do_betlejem',
-              'http://spiewnikreligijny.pl/teksty/2553,polnoc_juz_byla',
-              'http://spiewnikreligijny.pl/teksty/2219,pozno___juz____',
-              'http://spiewnikreligijny.pl/teksty/2997,pragne',
-              'http://spiewnikreligijny.pl/teksty/2325,pragne_tobie_dzis_wyspiewac___',
-              'http://spiewnikreligijny.pl/teksty/2822,pragne_z_toba_byc',
-              'http://spiewnikreligijny.pl/teksty/2692,prawda_jedyna',
-              'http://spiewnikreligijny.pl/teksty/1814,promyk_swiatla',
-              'http://spiewnikreligijny.pl/teksty/3248,prosba',
-              'http://spiewnikreligijny.pl/teksty/1320,prosby_do_boga',
-              'http://spiewnikreligijny.pl/teksty/2424,prowadz_gwiazdo',
-              'http://spiewnikreligijny.pl/teksty/2898,prowadz_mnie',
-              'http://spiewnikreligijny.pl/teksty/4306,prowadz_mnie',
-              'http://spiewnikreligijny.pl/teksty/4086,prowadz_mnie_do_chwaly',
-              'http://spiewnikreligijny.pl/teksty/210,prowadz_mnie_droga',
-              'http://spiewnikreligijny.pl/teksty/3003,prowadz_panie',
-              'http://spiewnikreligijny.pl/teksty/60,przebaczone_winy_darowane_dlugi',
-              'http://spiewnikreligijny.pl/teksty/2323,przebudzony_ranna_mgla',
-              'http://spiewnikreligijny.pl/teksty/619,przeciez_wiesz_ze_kocham',
-              'http://spiewnikreligijny.pl/teksty/2326,przed_obliczem_pana',
-              'http://spiewnikreligijny.pl/teksty/211,przed_oczy_twoje_panie',
-              'http://spiewnikreligijny.pl/teksty/3231,przed_tak_wielkim_sakramentem_',
-              'http://spiewnikreligijny.pl/teksty/3104,przed_wielu_laty',
-              'http://spiewnikreligijny.pl/teksty/1938,przedsionki',
-              'http://spiewnikreligijny.pl/teksty/512,przemien_nas_panie',
-              'http://spiewnikreligijny.pl/teksty/1037,przeogromna_ziemio',
-              'http://spiewnikreligijny.pl/teksty/2869,przepraszam',
-              'http://spiewnikreligijny.pl/teksty/1676,przepraszam_cie_boze',
-              'http://spiewnikreligijny.pl/teksty/3086,przepraszam_cie_dzisiaj_(akt_pokutny)',
-              'http://spiewnikreligijny.pl/teksty/212,przepusc_panie_przepusc_ludowi_twemu',
-              'http://spiewnikreligijny.pl/teksty/1704,przeslanie_rekolekcyjne',
-              'http://spiewnikreligijny.pl/teksty/1974,przez_meke_trzeba_isc',
-              'http://spiewnikreligijny.pl/teksty/2521,przez_wielkie_miasta',
-              'http://spiewnikreligijny.pl/teksty/1079,przy_onej_dolinie',
-              'http://spiewnikreligijny.pl/teksty/445,przy_onej_gorze',
-              'http://spiewnikreligijny.pl/teksty/4030,przybadz_duchu_swiety_',
-              'http://spiewnikreligijny.pl/teksty/42,przybadz_panie_bo_czekamy',
-              'http://spiewnikreligijny.pl/teksty/3027,przybadz_plomieniu',
-              'http://spiewnikreligijny.pl/teksty/443,przybiezeli_do_betlejem_pasterze',
-              'http://spiewnikreligijny.pl/teksty/2882,przychodze_znow_do_mego_pana',
-              'http://spiewnikreligijny.pl/teksty/4908,przychodzi_gosc',
-              'http://spiewnikreligijny.pl/teksty/1626,przygotuje_ci_serce_o_chryste_',
-              'http://spiewnikreligijny.pl/teksty/4810,przyjaciel',
-              'http://spiewnikreligijny.pl/teksty/943,przyjaciela_mam',
-              'http://spiewnikreligijny.pl/teksty/1137,przyjacielu',
-              'http://spiewnikreligijny.pl/teksty/1752,przyjacielu_drogi_stoj_to_chrystus_wola_cie',
-              'http://spiewnikreligijny.pl/teksty/3326,przyjdz',
-              'http://spiewnikreligijny.pl/teksty/2387,przyjdz_do_mnie_maly_jezusiczku',
-              'http://spiewnikreligijny.pl/teksty/2833,przyjdz_duchu_siwety',
-              'http://spiewnikreligijny.pl/teksty/2888,przyjdz_duchu_swiety_',
-              'http://spiewnikreligijny.pl/teksty/2927,przyjdz_dzieciatko',
-              'http://spiewnikreligijny.pl/teksty/2992,przyjdz_i_rozpal_nas',
-              'http://spiewnikreligijny.pl/teksty/2473,przyjdz_panie',
-              'http://spiewnikreligijny.pl/teksty/3038,przyjdz_z_poklonem_ludu_bozy',
-              'http://spiewnikreligijny.pl/teksty/2320,przyjdzcie_do_stolu_milosierdzia',
-              'http://spiewnikreligijny.pl/teksty/2327,przyjdzcie_radosnie',
-              'http://spiewnikreligijny.pl/teksty/798,przyjmij_o_najswietszy_panie',
-              'http://spiewnikreligijny.pl/teksty/706,przyjmij_ofiare_ojcze_nasz',
-              'http://spiewnikreligijny.pl/teksty/2686,przyjmij_panie',
-              'http://spiewnikreligijny.pl/teksty/705,przyjmij_panie_wolnosc_moja',
-              'http://spiewnikreligijny.pl/teksty/800,przykazanie',
-              'http://spiewnikreligijny.pl/teksty/444,przylecieli_aniolkowie',
-              'http://spiewnikreligijny.pl/teksty/3017,przylecieli_tak_sliczni',
-              'http://spiewnikreligijny.pl/teksty/799,przynosimy_ci_ojcze_swiety',
-              'http://spiewnikreligijny.pl/teksty/213,przystapcie_blizej_grzesznicy',
-              'http://spiewnikreligijny.pl/teksty/442,przystapmy_do_szopy_usciskajmy',
-              'http://spiewnikreligijny.pl/teksty/2382,przyszedl_na_swiat_',
-              'http://spiewnikreligijny.pl/teksty/1219,przytul_mnie_do_swoich_ran',
-              'http://spiewnikreligijny.pl/teksty/2810,psalm', 'http://spiewnikreligijny.pl/teksty/2314,psalm',
-              'http://spiewnikreligijny.pl/teksty/582,psalm_103', 'http://spiewnikreligijny.pl/teksty/3409,psalm_108',
-              'http://spiewnikreligijny.pl/teksty/600,psalm_122',
-              'http://spiewnikreligijny.pl/teksty/4743,psalm_125_(izrael_)',
-              'http://spiewnikreligijny.pl/teksty/579,psalm_13',
-              'http://spiewnikreligijny.pl/teksty/602,psalm_130_(z_glebokosci_wolam_do_ciebie)',
-              'http://spiewnikreligijny.pl/teksty/2711,psalm_145', 'http://spiewnikreligijny.pl/teksty/977,psalm_148',
-              'http://spiewnikreligijny.pl/teksty/583,psalm_18_(kocham_cie_panie)',
-              'http://spiewnikreligijny.pl/teksty/596,psalm_23', 'http://spiewnikreligijny.pl/teksty/3257,psalm_23_ii',
-              'http://spiewnikreligijny.pl/teksty/527,psalm_24', 'http://spiewnikreligijny.pl/teksty/950,psalm_29',
-              'http://spiewnikreligijny.pl/teksty/585,psalm_34', 'http://spiewnikreligijny.pl/teksty/593,psalm_40',
-              'http://spiewnikreligijny.pl/teksty/591,psalm_51_(zmiluj_sie_boze_nade_mna)',
-              'http://spiewnikreligijny.pl/teksty/3416,psalm_63',
-              'http://spiewnikreligijny.pl/teksty/568,psalm_8_(o_panie_nasz_panie)',
-              'http://spiewnikreligijny.pl/teksty/3552,ptaki_niby_dzwoneczki',
-              'http://spiewnikreligijny.pl/teksty/1151,ptaki_radosne_niestrudzone',
-              'http://spiewnikreligijny.pl/teksty/2977,puk_puk',
-              'http://spiewnikreligijny.pl/teksty/3280,puk_puk_puk_(oto_ja)',
-              'http://spiewnikreligijny.pl/teksty/612,pustynia', 'http://spiewnikreligijny.pl/teksty/208,pustynia',
-              'http://spiewnikreligijny.pl/teksty/209,pustynia_jest_laska',
-              'http://spiewnikreligijny.pl/teksty/2034,quo_vadis',
-              'http://spiewnikreligijny.pl/teksty/801,racz_wiekuiste',
-              'http://spiewnikreligijny.pl/teksty/802,radio_maryja_sluchamy',
-              'http://spiewnikreligijny.pl/teksty/1800,radosny_dzien',
-              'http://spiewnikreligijny.pl/teksty/544,radosny_jestem', 'http://spiewnikreligijny.pl/teksty/803,radosc',
-              'http://spiewnikreligijny.pl/teksty/2328,radosc_tchnij',
-              'http://spiewnikreligijny.pl/teksty/449,raduj_sie_swiecie',
-              'http://spiewnikreligijny.pl/teksty/2510,raduj_sie_',
-              'http://spiewnikreligijny.pl/teksty/22,raduj_sie_nieba_krolowo',
-              'http://spiewnikreligijny.pl/teksty/446,raduj_sie_ziemio',
-              'http://spiewnikreligijny.pl/teksty/804,raduj_sie_chwal_imie_pana',
-              'http://spiewnikreligijny.pl/teksty/1904,raduje_sie_dusza_ma',
-              'http://spiewnikreligijny.pl/teksty/4811,radujmy_sie',
-              'http://spiewnikreligijny.pl/teksty/447,radujmy_sie_bracia_mili',
-              'http://spiewnikreligijny.pl/teksty/3267,ramiona_szeroko_otwarte',
-              'http://spiewnikreligijny.pl/teksty/4818,rastafarianty_(sny)',
-              'http://spiewnikreligijny.pl/teksty/75,ratuj_mnie', 'http://spiewnikreligijny.pl/teksty/3224,ratuj_panie',
-              'http://spiewnikreligijny.pl/teksty/805,raz_dwa_trzy',
-              'http://spiewnikreligijny.pl/teksty/3402,redemption_song', 'http://spiewnikreligijny.pl/teksty/1858,rejs',
-              'http://spiewnikreligijny.pl/teksty/4760,robaczywe_jablka',
-              'http://spiewnikreligijny.pl/teksty/4758,roboty_do_roboty',
-              'http://spiewnikreligijny.pl/teksty/4778,rodzi_sie_bog',
-              'http://spiewnikreligijny.pl/teksty/1699,rodzinka',
-              'http://spiewnikreligijny.pl/teksty/774,rodzino_radia_maryja',
-              'http://spiewnikreligijny.pl/teksty/806,rosnie_chleb',
-              'http://spiewnikreligijny.pl/teksty/448,rozkwitnela_sie_lilija',
-              'http://spiewnikreligijny.pl/teksty/217,rozmowa',
-              'http://spiewnikreligijny.pl/teksty/215,rozmyslajmy_dzis_wierni_chrzescijanie',
-              'http://spiewnikreligijny.pl/teksty/4734,rozpedzony',
-              'http://spiewnikreligijny.pl/teksty/214,rozpiety_na_ramionach',
-              'http://spiewnikreligijny.pl/teksty/216,rozplyncie_sie_me_zrenice',
-              'http://spiewnikreligijny.pl/teksty/3344,rsr', 'http://spiewnikreligijny.pl/teksty/807,ruah',
-              'http://spiewnikreligijny.pl/teksty/4740,runal_juz_ostatni_mur',
-              'http://spiewnikreligijny.pl/teksty/808,ruszaj_tam',
-              'http://spiewnikreligijny.pl/teksty/3237,ruszaj_ziarno',
-              'http://spiewnikreligijny.pl/teksty/2471,rysuje_krzyz', 'http://spiewnikreligijny.pl/teksty/809,rzeki',
-              'http://spiewnikreligijny.pl/teksty/3212,rzuc_mnie_w_morze',
-              'http://spiewnikreligijny.pl/teksty/3067,salve_rex_christe',
-              'http://spiewnikreligijny.pl/teksty/1238,same_sie_musza_kruszyc_twarde_skaly_',
-              'http://spiewnikreligijny.pl/teksty/1651,san_damiano', 'http://spiewnikreligijny.pl/teksty/3261,sanctus',
-              'http://spiewnikreligijny.pl/teksty/941,sandaly_(gdy_uczniow_swych_posylal_pan)',
-              'http://spiewnikreligijny.pl/teksty/4759,santi_santi',
-              'http://spiewnikreligijny.pl/teksty/2905,santo_subito',
-              'http://spiewnikreligijny.pl/teksty/2201,schemma_izrael',
-              'http://spiewnikreligijny.pl/teksty/4848,schowaj_mnie',
-              'http://spiewnikreligijny.pl/teksty/3397,see_i_and_i',
-              'http://spiewnikreligijny.pl/teksty/2729,sekwencja_do_ducha_swietego',
-              'http://spiewnikreligijny.pl/teksty/3186,serce_boga',
-              'http://spiewnikreligijny.pl/teksty/3441,serce_jezusa_krolem_polski',
-              'http://spiewnikreligijny.pl/teksty/225,serce_me_do_ciebie_wznosze',
-              'http://spiewnikreligijny.pl/teksty/2329,serce_moje_jest_mocne',
-              'http://spiewnikreligijny.pl/teksty/3419,serce_oddaje_ci',
-              'http://spiewnikreligijny.pl/teksty/4834,sezam', 'http://spiewnikreligijny.pl/teksty/592,shalom',
-              'http://spiewnikreligijny.pl/teksty/576,shema_israel_adonai_elohenu_adonai_ehad',
-              'http://spiewnikreligijny.pl/teksty/580,shlom_lech_mariam',
-              'http://spiewnikreligijny.pl/teksty/65,sieje_je',
-              'http://spiewnikreligijny.pl/teksty/4898,sierpniowe_anioly',
-              'http://spiewnikreligijny.pl/teksty/3333,siubi_dubi_(psalm_88)',
-              'http://spiewnikreligijny.pl/teksty/3283,skala', 'http://spiewnikreligijny.pl/teksty/3191,skoro_panie',
-              'http://spiewnikreligijny.pl/teksty/3288,skrzydlaty',
-              'http://spiewnikreligijny.pl/teksty/450,skrzypi_woz',
-              'http://spiewnikreligijny.pl/teksty/4738,skuter_andrzeja',
-              'http://spiewnikreligijny.pl/teksty/3089,slaw_jezyku_tajemnice',
-              'http://spiewnikreligijny.pl/teksty/2774,slawie_panie_imie_twe',
-              'http://spiewnikreligijny.pl/teksty/3213,slawmy_pana',
-              'http://spiewnikreligijny.pl/teksty/4838,slodkie_slowa',
-              'http://spiewnikreligijny.pl/teksty/3370,slonce_nagle_zgaslo',
-              'http://spiewnikreligijny.pl/teksty/95,slonce_zza_chmury',
-              'http://spiewnikreligijny.pl/teksty/3215,slowa_twych_ust',
-              'http://spiewnikreligijny.pl/teksty/3265,slowianski', 'http://spiewnikreligijny.pl/teksty/3407,slowo',
-              'http://spiewnikreligijny.pl/teksty/4827,slowo_cialem',
-              'http://spiewnikreligijny.pl/teksty/2330,sluchaj_izraelu',
-              'http://spiewnikreligijny.pl/teksty/3434,sluchaj_izraelu',
-              'http://spiewnikreligijny.pl/teksty/574,sluchaj_izraelu_(jedyny_pan)',
-              'http://spiewnikreligijny.pl/teksty/3369,sluchaj_izraelu_(pan_jest_twoim_bogiem)',
-              'http://spiewnikreligijny.pl/teksty/2331,slysze_juz_armii_pana_glos',
-              'http://spiewnikreligijny.pl/teksty/451,slysze_z_nieba_muzyke',
-              'http://spiewnikreligijny.pl/teksty/1688,sos', 'http://spiewnikreligijny.pl/teksty/515,sol_swiata',
-              'http://spiewnikreligijny.pl/teksty/4799,spacer(ek)',
-              'http://spiewnikreligijny.pl/teksty/1804,spelnione_marzenia',
-              'http://spiewnikreligijny.pl/teksty/2332,spocznij_na_nas',
-              'http://spiewnikreligijny.pl/teksty/4108,spojrzal_jezus_na_szymona',
-              'http://spiewnikreligijny.pl/teksty/1209,spojrzenia_twe_',
-              'http://spiewnikreligijny.pl/teksty/2826,spojrzyj_moj_boze',
-              'http://spiewnikreligijny.pl/teksty/223,spojrzyj_z_nieba_wysokiego',
-              'http://spiewnikreligijny.pl/teksty/2333,spotkal_mnie_dzis_pan',
-              'http://spiewnikreligijny.pl/teksty/3313,spotkalem_dzis_jezusa',
-              'http://spiewnikreligijny.pl/teksty/517,spotkanie',
-              'http://spiewnikreligijny.pl/teksty/1169,spotkanie_z_ukrzyzowanym',
-              'http://spiewnikreligijny.pl/teksty/2853,spojrz_boze_mocny',
-              'http://spiewnikreligijny.pl/teksty/3106,spojrz_na_nasze_dary',
-              'http://spiewnikreligijny.pl/teksty/810,spojrz_tam_w_gorze',
-              'http://spiewnikreligijny.pl/teksty/3053,spraw_maryjo',
-              'http://spiewnikreligijny.pl/teksty/4889,sprawa_aniola',
-              'http://spiewnikreligijny.pl/teksty/1227,sprawie_ze_',
-              'http://spiewnikreligijny.pl/teksty/44,spuscie_rose_niebiosa_z_gory',
-              'http://spiewnikreligijny.pl/teksty/43,spusccie_nam_na_ziemskie_niwy',
-              'http://spiewnikreligijny.pl/teksty/4742,spytaj_milicjanta',
-              'http://spiewnikreligijny.pl/teksty/2400,srebra_i_zlota',
-              'http://spiewnikreligijny.pl/teksty/1019,srebrna_gwiazdka',
-              'http://spiewnikreligijny.pl/teksty/588,stabat_mater',
-              'http://spiewnikreligijny.pl/teksty/4855,stale_glodna',
-              'http://spiewnikreligijny.pl/teksty/224,stal_sie_symbolem_z_drzazgami',
-              'http://spiewnikreligijny.pl/teksty/3372,stala_matka_bolesciwa',
-              'http://spiewnikreligijny.pl/teksty/221,stala_matka_bolesciwa_ii',
-              'http://spiewnikreligijny.pl/teksty/220,stane_pod_krzyzem',
-              'http://spiewnikreligijny.pl/teksty/1043,staniesz_tak_boze',
-              'http://spiewnikreligijny.pl/teksty/219,stan_pod_krzyzem',
-              'http://spiewnikreligijny.pl/teksty/1948,startuj', 'http://spiewnikreligijny.pl/teksty/81,stary_abraham',
-              'http://spiewnikreligijny.pl/teksty/2718,stary_krzyz',
-              'http://spiewnikreligijny.pl/teksty/4761,stary_szawel_nowy_pawel',
-              'http://spiewnikreligijny.pl/teksty/4736,stroz', 'http://spiewnikreligijny.pl/teksty/562,stroz_poranka',
-              'http://spiewnikreligijny.pl/teksty/4888,strych_z_aniolami',
-              'http://spiewnikreligijny.pl/teksty/45,stworzycielu_gwiazd_swiecacych',
-              'http://spiewnikreligijny.pl/teksty/3331,sunshine',
-              'http://spiewnikreligijny.pl/teksty/903,sureksit_christus',
-              'http://spiewnikreligijny.pl/teksty/4868,sursum_corda', 'http://spiewnikreligijny.pl/teksty/4110,sutanna',
-              'http://spiewnikreligijny.pl/teksty/4109,syn_marnotrawny',
-              'http://spiewnikreligijny.pl/teksty/3139,syn_sluzebnicy',
-              'http://spiewnikreligijny.pl/teksty/1838,synu_dawida',
-              'http://spiewnikreligijny.pl/teksty/2319,szachownica', 'http://spiewnikreligijny.pl/teksty/1161,szanta',
-              'http://spiewnikreligijny.pl/teksty/1168,szara_droga',
-              'http://spiewnikreligijny.pl/teksty/2750,szczescie',
-              'http://spiewnikreligijny.pl/teksty/2385,szczesliwi_mlodzi',
-              'http://spiewnikreligijny.pl/teksty/1117,szczesliwy_pastuszek',
-              'http://spiewnikreligijny.pl/teksty/2338,szedlem_kiedys_sciezyna',
-              'http://spiewnikreligijny.pl/teksty/594,szema_izrael',
-              'http://spiewnikreligijny.pl/teksty/4775,szema_izrael',
-              'http://spiewnikreligijny.pl/teksty/3349,szescset_lat_maryjo',
-              'http://spiewnikreligijny.pl/teksty/1268,szukac_warto',
-              'http://spiewnikreligijny.pl/teksty/811,szukajcie_wpierw_krolestwa_bozego',
-              'http://spiewnikreligijny.pl/teksty/4774,szukalem_milosci',
-              'http://spiewnikreligijny.pl/teksty/2340,szukam_mojego_pana',
-              'http://spiewnikreligijny.pl/teksty/812,szukam_sensu_zycia',
-              'http://spiewnikreligijny.pl/teksty/2339,szumia_w_lesie_drzewa',
-              'http://spiewnikreligijny.pl/teksty/222,szymonie_cyrenejczyku',
-              'http://spiewnikreligijny.pl/teksty/1167,sciezka_przez_las',
-              'http://spiewnikreligijny.pl/teksty/908,slady',
-              'http://spiewnikreligijny.pl/teksty/454,sliczna_gwiazda_wschodzi_nam',
-              'http://spiewnikreligijny.pl/teksty/453,sliczna_panienka_jezusa_zrodzila',
-              'http://spiewnikreligijny.pl/teksty/452,sliczna_panienka_jezusa_zrodzila_i',
-              'http://spiewnikreligijny.pl/teksty/218,sliczny_jezu_mily_panie',
-              'http://spiewnikreligijny.pl/teksty/2961,smialy_badz_mezny_badz',
-              'http://spiewnikreligijny.pl/teksty/4869,smierc_w_grobie',
-              'http://spiewnikreligijny.pl/teksty/3228,spiewa_izrael',
-              'http://spiewnikreligijny.pl/teksty/2849,spiewa_mi_slowiczek',
-              'http://spiewnikreligijny.pl/teksty/2934,spiewaj_panu',
-              'http://spiewnikreligijny.pl/teksty/3391,spiewajmy_alleluja_barankowi',
-              'http://spiewnikreligijny.pl/teksty/3209,spiewam_tobie_magnificat',
-              'http://spiewnikreligijny.pl/teksty/2341,spiewam_tobie_piesn_chwaly',
-              'http://spiewnikreligijny.pl/teksty/3440,spij_malenki_jezu',
-              'http://spiewnikreligijny.pl/teksty/2342,swiat_bylby_taki_ubogi',
-              'http://spiewnikreligijny.pl/teksty/978,swiat_marzen',
-              'http://spiewnikreligijny.pl/teksty/813,swiat_nie_jest_domem_mym',
-              'http://spiewnikreligijny.pl/teksty/814,swiatlem_i_zbawieniem_moim_pan',
-              'http://spiewnikreligijny.pl/teksty/2590,swiatlem_i_zbawieniem_mym',
-              'http://spiewnikreligijny.pl/teksty/3347,swiatlo',
-              'http://spiewnikreligijny.pl/teksty/3605,swiatlosci_swiata',
-              'http://spiewnikreligijny.pl/teksty/1600,swiatlosc',
-              'http://spiewnikreligijny.pl/teksty/815,swiatlosc_swiata',
-              'http://spiewnikreligijny.pl/teksty/929,swieczka',
-              'http://spiewnikreligijny.pl/teksty/3466,swiec_gwiazdeczko_swiec',
-              'http://spiewnikreligijny.pl/teksty/816,swiec_jezu_swiec',
-              'http://spiewnikreligijny.pl/teksty/2901,swieta_maryjo_krolowo',
-              'http://spiewnikreligijny.pl/teksty/975,swieta_rodzina',
-              'http://spiewnikreligijny.pl/teksty/892,swiete_imie_jezus',
-              'http://spiewnikreligijny.pl/teksty/1810,swiete_miejsca',
-              'http://spiewnikreligijny.pl/teksty/2240,swietemu_bogu_oddaj_czesc',
-              'http://spiewnikreligijny.pl/teksty/3374,swiety_boze',
-              'http://spiewnikreligijny.pl/teksty/3281,swiety_spokoj',
-              'http://spiewnikreligijny.pl/teksty/3404,swiety_szczyt',
-              'http://spiewnikreligijny.pl/teksty/3188,swiety_swiety_(greckie)',
-              'http://spiewnikreligijny.pl/teksty/59,swiety_swiety_usmiechniety',
-              'http://spiewnikreligijny.pl/teksty/1107,swinty_scepon',
-              'http://spiewnikreligijny.pl/teksty/1177,ta_krew',
-              'http://spiewnikreligijny.pl/teksty/1171,tacy_sami_jak_my',
-              'http://spiewnikreligijny.pl/teksty/552,tak_blisko_jestes_wszystkich',
-              'http://spiewnikreligijny.pl/teksty/533,tak_cudowna',
-              'http://spiewnikreligijny.pl/teksty/1175,tak_jak_maryja',
-              'http://spiewnikreligijny.pl/teksty/227,tak_mnie_skrusz',
-              'http://spiewnikreligijny.pl/teksty/230,tak_mowi_amen',
-              'http://spiewnikreligijny.pl/teksty/1860,tak_tak_panie',
-              'http://spiewnikreligijny.pl/teksty/539,tak_tak_wyglada_moj_swiat',
-              'http://spiewnikreligijny.pl/teksty/4887,tak_to_jest',
-              'http://spiewnikreligijny.pl/teksty/4755,tak_to_tak_a_nie_to_nie',
-              'http://spiewnikreligijny.pl/teksty/1961,tak_wyschles_panie',
-              'http://spiewnikreligijny.pl/teksty/1498,tak_tak_panie',
-              'http://spiewnikreligijny.pl/teksty/1983,tak_tak_tak_o_panie_moj',
-              'http://spiewnikreligijny.pl/teksty/4881,taka_sobie_piosenka',
-              'http://spiewnikreligijny.pl/teksty/2343,taki_duzy_taki_maly',
-              'http://spiewnikreligijny.pl/teksty/615,takiego_nieba',
-              'http://spiewnikreligijny.pl/teksty/4824,takiej_milosci_szukales',
-              'http://spiewnikreligijny.pl/teksty/1524,tam_gdzie_milosc_jest',
-              'http://spiewnikreligijny.pl/teksty/518,tancem_chwalmy_go',
-              'http://spiewnikreligijny.pl/teksty/538,tanczymy_noca',
-              'http://spiewnikreligijny.pl/teksty/1794,tasiemka', 'http://spiewnikreligijny.pl/teksty/4900,tata',
-              'http://spiewnikreligijny.pl/teksty/57,tato_(nie_boje_sie_gdy_ciemno_jest)',
-              'http://spiewnikreligijny.pl/teksty/897,tchnij_moc',
-              'http://spiewnikreligijny.pl/teksty/2519,tej_nocy_aniol_przyszedl_do_jozefa',
-              'http://spiewnikreligijny.pl/teksty/3241,telefon',
-              'http://spiewnikreligijny.pl/teksty/228,ten_ktorego_pocaluje_w_twarz',
-              'http://spiewnikreligijny.pl/teksty/542,ten_obcy',
-              'http://spiewnikreligijny.pl/teksty/2366,ten_szczegolny_dzien_sie_budzi',
-              'http://spiewnikreligijny.pl/teksty/456,teraz_spij_dziecino_mala',
-              'http://spiewnikreligijny.pl/teksty/2184,tlumy_serc',
-              'http://spiewnikreligijny.pl/teksty/1028,to_maryja_moja_mama',
-              'http://spiewnikreligijny.pl/teksty/2250,to_moj_pan',
-              'http://spiewnikreligijny.pl/teksty/1499,to_moj_pan_wiec_nie_lekam_sie',
-              'http://spiewnikreligijny.pl/teksty/229,to_nie_gwozdzie_cie_przebily',
-              'http://spiewnikreligijny.pl/teksty/1861,to_przykazanie',
-              'http://spiewnikreligijny.pl/teksty/4756,to_sie_dzisiaj_stanie',
-              'http://spiewnikreligijny.pl/teksty/2488,to_swiat_czas',
-              'http://spiewnikreligijny.pl/teksty/1761,to_ty_wziales_mnie_w_ramiona',
-              'http://spiewnikreligijny.pl/teksty/2312,to_tylko_dzis',
-              'http://spiewnikreligijny.pl/teksty/613,to_wlasnie_zycie_jest',
-              'http://spiewnikreligijny.pl/teksty/862,tobie_chor_aniolow_spiewa',
-              'http://spiewnikreligijny.pl/teksty/46,tobie_nad_pomysl_piekno_i_wymowe',
-              'http://spiewnikreligijny.pl/teksty/2845,tobie_panie_zaufalem',
-              'http://spiewnikreligijny.pl/teksty/3095,tobie_ufam', 'http://spiewnikreligijny.pl/teksty/818,tolerancja',
-              'http://spiewnikreligijny.pl/teksty/455,triumfy_krola_niebieskiego',
-              'http://spiewnikreligijny.pl/teksty/3143,trojco_swieta',
-              'http://spiewnikreligijny.pl/teksty/1862,trudna_milosc',
-              'http://spiewnikreligijny.pl/teksty/3328,trust_(psalm_27)',
-              'http://spiewnikreligijny.pl/teksty/2453,trzeba_isc', 'http://spiewnikreligijny.pl/teksty/2346,trzoda',
-              'http://spiewnikreligijny.pl/teksty/2739,trzy_maryje',
-              'http://spiewnikreligijny.pl/teksty/4885,trzy_pytania',
-              'http://spiewnikreligijny.pl/teksty/3298,trzy_slonca',
-              'http://spiewnikreligijny.pl/teksty/3799,tu_jest_miejsce_utajenia',
-              'http://spiewnikreligijny.pl/teksty/3245,tup_tup_puk_puk',
-              'http://spiewnikreligijny.pl/teksty/894,twe_drogi',
-              'http://spiewnikreligijny.pl/teksty/1956,twe_drogi_sa_najlepsze',
-              'http://spiewnikreligijny.pl/teksty/2344,twe_milosierdzie',
-              'http://spiewnikreligijny.pl/teksty/2721,twe_swiatlo_jest',
-              'http://spiewnikreligijny.pl/teksty/2269,twoja_krew',
-              'http://spiewnikreligijny.pl/teksty/1144,twoja_milosc',
-              'http://spiewnikreligijny.pl/teksty/4783,twoja_milosc',
-              'http://spiewnikreligijny.pl/teksty/2345,twoja_milosc',
-              'http://spiewnikreligijny.pl/teksty/1955,twoja_wola', 'http://spiewnikreligijny.pl/teksty/561,twoje_dary',
-              'http://spiewnikreligijny.pl/teksty/2528,twoje_fiat',
-              'http://spiewnikreligijny.pl/teksty/604,twoje_milowanie',
-              'http://spiewnikreligijny.pl/teksty/1803,twoje_piec_minut',
-              'http://spiewnikreligijny.pl/teksty/3386,twoje_slowo',
-              'http://spiewnikreligijny.pl/teksty/77,twoje_swiatlo',
-              'http://spiewnikreligijny.pl/teksty/2705,twoj_krzyz',
-              'http://spiewnikreligijny.pl/teksty/819,ty_dasz_mi_pokoj_serca',
-              'http://spiewnikreligijny.pl/teksty/870,ty_jestes_skala_zbawienia_mego',
-              'http://spiewnikreligijny.pl/teksty/1718,ty_panie',
-              'http://spiewnikreligijny.pl/teksty/3468,ty_panie_i_tylko_ty',
-              'http://spiewnikreligijny.pl/teksty/4782,ty_patrzysz_z_gory',
-              'http://spiewnikreligijny.pl/teksty/2389,ty_swiatlosc_dnia',
-              'http://spiewnikreligijny.pl/teksty/532,ty_tylko_mnie_poprowadz',
-              'http://spiewnikreligijny.pl/teksty/4745,ty_tylko_ty_(brygada_kryzys)',
-              'http://spiewnikreligijny.pl/teksty/3714,ty_wladaj',
-              'http://spiewnikreligijny.pl/teksty/231,ty_cos_synem_bozym_jest',
-              'http://spiewnikreligijny.pl/teksty/226,ty_ktorys_gorzko_na_krzyzu_umieral',
-              'http://spiewnikreligijny.pl/teksty/2197,tyle_dobrego',
-              'http://spiewnikreligijny.pl/teksty/550,tylko_badz', 'http://spiewnikreligijny.pl/teksty/4116,tylko_bog',
-              'http://spiewnikreligijny.pl/teksty/3297,tylko_brak_mi_czasu',
-              'http://spiewnikreligijny.pl/teksty/4800,tylko_jezus', 'http://spiewnikreligijny.pl/teksty/3341,tylko_on',
-              'http://spiewnikreligijny.pl/teksty/1163,tylko_orly',
-              'http://spiewnikreligijny.pl/teksty/1696,tylko_rodzinka',
-              'http://spiewnikreligijny.pl/teksty/820,tylko_slowo',
-              'http://spiewnikreligijny.pl/teksty/531,tylko_troche_trzeba_dac',
-              'http://spiewnikreligijny.pl/teksty/2347,tylko_w_nim_jest_zbawienie',
-              'http://spiewnikreligijny.pl/teksty/2968,tylko_w_twoim_milosierdziu',
-              'http://spiewnikreligijny.pl/teksty/2823,tys_dla_mnie_ucieczka',
-              'http://spiewnikreligijny.pl/teksty/3752,tys_jak_powietrza_wiew',
-              'http://spiewnikreligijny.pl/teksty/821,tys_jak_skala_tys_jak_wzgorze',
-              'http://spiewnikreligijny.pl/teksty/3586,tys_jest_bog__',
-              'http://spiewnikreligijny.pl/teksty/2384,tys_jest_swiety',
-              'http://spiewnikreligijny.pl/teksty/3146,tys_nasza_moca',
-              'http://spiewnikreligijny.pl/teksty/3365,tys_w_wieczerniku',
-              'http://spiewnikreligijny.pl/teksty/827,u_drzwi_twoich_stoje_panie',
-              'http://spiewnikreligijny.pl/teksty/2797,u_pana', 'http://spiewnikreligijny.pl/teksty/822,ubi_caritas',
-              'http://spiewnikreligijny.pl/teksty/2310,uboga_stejenka',
-              'http://spiewnikreligijny.pl/teksty/457,ucieczka_do_egiptu',
-              'http://spiewnikreligijny.pl/teksty/2567,uciekali', 'http://spiewnikreligijny.pl/teksty/823,uciszcie_sie',
-              'http://spiewnikreligijny.pl/teksty/2949,uczniowie_pana', 'http://spiewnikreligijny.pl/teksty/3114,uczta',
-              'http://spiewnikreligijny.pl/teksty/1093,uczyl_nas',
-              'http://spiewnikreligijny.pl/teksty/825,uczyncie_co_wam_mowi_syn',
-              'http://spiewnikreligijny.pl/teksty/826,uderza_na_miasto',
-              'http://spiewnikreligijny.pl/teksty/3294,ufam_tobie',
-              'http://spiewnikreligijny.pl/teksty/2726,ujalem_cie_za_prawice',
-              'http://spiewnikreligijny.pl/teksty/519,ukaz_mi_panie_swa_twarz',
-              'http://spiewnikreligijny.pl/teksty/710,ukochales_prosty_habit',
-              'http://spiewnikreligijny.pl/teksty/3205,ukrzyzowana_milosc',
-              'http://spiewnikreligijny.pl/teksty/1697,ulecz_swiat',
-              'http://spiewnikreligijny.pl/teksty/828,upadnij_na_kolana',
-              'http://spiewnikreligijny.pl/teksty/47,urzad_zbawienia_ludzkiego',
-              'http://spiewnikreligijny.pl/teksty/824,uslysz_bozej_matki_glos',
-              'http://spiewnikreligijny.pl/teksty/2472,uslysz_dobry_boze',
-              'http://spiewnikreligijny.pl/teksty/232,uslysz_panie_dobrotliwy',
-              'http://spiewnikreligijny.pl/teksty/670,uslyszalem_jakby_glos',
-              'http://spiewnikreligijny.pl/teksty/3246,uslyszales_szept',
-              'http://spiewnikreligijny.pl/teksty/829,usmiech',
-              'http://spiewnikreligijny.pl/teksty/2984,usmiechnij_sie',
-              'http://spiewnikreligijny.pl/teksty/830,uwielbia_ciebie_ziemia',
-              'http://spiewnikreligijny.pl/teksty/831,uwielbia_dusza_moja_pana',
-              'http://spiewnikreligijny.pl/teksty/871,uwielbiac_mego_pana_chce',
-              'http://spiewnikreligijny.pl/teksty/836,uwielbiajcie_pana',
-              'http://spiewnikreligijny.pl/teksty/835,uwielbiajmy_chrystusa',
-              'http://spiewnikreligijny.pl/teksty/91,uwielbiam_cie',
-              'http://spiewnikreligijny.pl/teksty/833,uwielbiam_cie_jezu',
-              'http://spiewnikreligijny.pl/teksty/832,uwielbiam_cie_blogoslawie_cie',
-              'http://spiewnikreligijny.pl/teksty/837,uwielbiam_imie_twoje_panie',
-              'http://spiewnikreligijny.pl/teksty/938,uwielbij_go_zaspiewaj_hymn',
-              'http://spiewnikreligijny.pl/teksty/571,uwiodles_mnie_panie',
-              'http://spiewnikreligijny.pl/teksty/838,veni_creator',
-              'http://spiewnikreligijny.pl/teksty/520,veni_spiritus_sanctus',
-              'http://spiewnikreligijny.pl/teksty/839,veritatis_splendor',
-              'http://spiewnikreligijny.pl/teksty/840,vox_nostra', 'http://spiewnikreligijny.pl/teksty/4903,w_betlejem',
-              'http://spiewnikreligijny.pl/teksty/467,w_boze_narodzenie_wszystko_sie_weseli',
-              'http://spiewnikreligijny.pl/teksty/318,w_cicha_noc_w_sniezna_noc',
-              'http://spiewnikreligijny.pl/teksty/1103,w_cieniu_twoich_rak',
-              'http://spiewnikreligijny.pl/teksty/48,w_czasie_swietym_adwentowym',
-              'http://spiewnikreligijny.pl/teksty/236,w_dialogu_milosci_trwac_chcemy',
-              'http://spiewnikreligijny.pl/teksty/464,w_dzien_bozego_narodzenia',
-              'http://spiewnikreligijny.pl/teksty/1061,w_grudniowe_noce',
-              'http://spiewnikreligijny.pl/teksty/2974,w_kazdym_strapieniu',
-              'http://spiewnikreligijny.pl/teksty/2549,w_krainie_zycia',
-              'http://spiewnikreligijny.pl/teksty/1637,w_kruszynie_chleba_',
-              'http://spiewnikreligijny.pl/teksty/234,w_krzyzu_cierpienie',
-              'http://spiewnikreligijny.pl/teksty/844,w_lekkim_powiewie',
-              'http://spiewnikreligijny.pl/teksty/1694,w_malenkiej_stajence',
-              'http://spiewnikreligijny.pl/teksty/1115,w_malenkiej_stajence',
-              'http://spiewnikreligijny.pl/teksty/466,w_najgorsza_z_zim',
-              'http://spiewnikreligijny.pl/teksty/468,w_narodzin_bozych_swiety_dzien',
-              'http://spiewnikreligijny.pl/teksty/843,w_nowe_jutro',
-              'http://spiewnikreligijny.pl/teksty/577,w_obliczu_aniolow',
-              'http://spiewnikreligijny.pl/teksty/235,w_ogrojcu_jezus_kleczy',
-              'http://spiewnikreligijny.pl/teksty/603,w_pismach_dla_pan',
-              'http://spiewnikreligijny.pl/teksty/473,w_slowie_jest_nasze_zycie',
-              'http://spiewnikreligijny.pl/teksty/1051,w_stajni',
-              'http://spiewnikreligijny.pl/teksty/1510,w_swe_ramiona_mnie_wez',
-              'http://spiewnikreligijny.pl/teksty/1017,w_swym_sercu_na_wieki',
-              'http://spiewnikreligijny.pl/teksty/3332,w_tobie_i_we_mnie',
-              'http://spiewnikreligijny.pl/teksty/1617,w_tobie_jest_swiatlo',
-              'http://spiewnikreligijny.pl/teksty/841,w_tobie_nadzieja',
-              'http://spiewnikreligijny.pl/teksty/1917,w_tobie_sila',
-              'http://spiewnikreligijny.pl/teksty/462,w_zlobie_lezy',
-              'http://spiewnikreligijny.pl/teksty/469,w_zlobku_na_sianie',
-              'http://spiewnikreligijny.pl/teksty/74,wakacje_caly_rok',
-              'http://spiewnikreligijny.pl/teksty/3346,warrior',
-              'http://spiewnikreligijny.pl/teksty/2146,warto_dla_jednej_milosci_zyc',
-              'http://spiewnikreligijny.pl/teksty/4858,waska_droga',
-              'http://spiewnikreligijny.pl/teksty/528,wciaz_pytam',
-              'http://spiewnikreligijny.pl/teksty/1962,wdziecznosc',
-              'http://spiewnikreligijny.pl/teksty/842,wejdzmy_do_jego_bram_z_dziekczynieniem',
-              'http://spiewnikreligijny.pl/teksty/964,weronika',
-              'http://spiewnikreligijny.pl/teksty/1256,wesel_sie_krolowo_mila',
-              'http://spiewnikreligijny.pl/teksty/28,wesel_sie_krolowo_rajska',
-              'http://spiewnikreligijny.pl/teksty/27,wesel_sie_ludu_strapiony',
-              'http://spiewnikreligijny.pl/teksty/26,wesel_sie_nieba_krolowo',
-              'http://spiewnikreligijny.pl/teksty/3439,wesel_sie_ziemio',
-              'http://spiewnikreligijny.pl/teksty/458,wesola_nowine',
-              'http://spiewnikreligijny.pl/teksty/7,wesoly_nam_dzien_dzis_nastal',
-              'http://spiewnikreligijny.pl/teksty/92,wez_mnie_na_rece',
-              'http://spiewnikreligijny.pl/teksty/1995,wez_moje_zycie',
-              'http://spiewnikreligijny.pl/teksty/845,wez_w_swa_opieke_nasz_kosciol_swiety',
-              'http://spiewnikreligijny.pl/teksty/4865,wedrowiec', 'http://spiewnikreligijny.pl/teksty/529,wedrujac',
-              'http://spiewnikreligijny.pl/teksty/4837,wiara_nadzieja_milosc',
-              'http://spiewnikreligijny.pl/teksty/3201,widze_nowe_niebo_i_ziemie',
-              'http://spiewnikreligijny.pl/teksty/847,wielbic_imie_pana_dzis_chce',
-              'http://spiewnikreligijny.pl/teksty/1181,wielbic_pana_chce',
-              'http://spiewnikreligijny.pl/teksty/1604,wielbie_ciebie_panie',
-              'http://spiewnikreligijny.pl/teksty/1597,wielbie_cie',
-              'http://spiewnikreligijny.pl/teksty/2348,wielbij_go',
-              'http://spiewnikreligijny.pl/teksty/1154,wielbij_pana',
-              'http://spiewnikreligijny.pl/teksty/2349,wiele_jest_serc',
-              'http://spiewnikreligijny.pl/teksty/4872,wielka_wygrana',
-              'http://spiewnikreligijny.pl/teksty/3422,wielki_jest_pan',
-              'http://spiewnikreligijny.pl/teksty/1566,wielkie_swiatlo',
-              'http://spiewnikreligijny.pl/teksty/4786,wiem_jestes',
-              'http://spiewnikreligijny.pl/teksty/4813,wiem_jestes_(psalm_63)',
-              'http://spiewnikreligijny.pl/teksty/2016,wierze_ci',
-              'http://spiewnikreligijny.pl/teksty/3064,wierze_w_ciebie_panie',
-              'http://spiewnikreligijny.pl/teksty/2470,wierze_w_ciebie_panie',
-              'http://spiewnikreligijny.pl/teksty/3012,wierze_ze_droga_ktora_ide',
-              'http://spiewnikreligijny.pl/teksty/238,wierzymy_w_ciebie_chryste',
-              'http://spiewnikreligijny.pl/teksty/848,wiecierz', 'http://spiewnikreligijny.pl/teksty/463,wigilia',
-              'http://spiewnikreligijny.pl/teksty/3254,wina_w_kazdym_z_nas',
-              'http://spiewnikreligijny.pl/teksty/233,wisi_na_krzyzu',
-              'http://spiewnikreligijny.pl/teksty/459,witaj_jezu_ukochany',
-              'http://spiewnikreligijny.pl/teksty/2661,witaj_krynico',
-              'http://spiewnikreligijny.pl/teksty/2800,witaj_majowa_jutrzenko',
-              'http://spiewnikreligijny.pl/teksty/2268,witaj_panie',
-              'http://spiewnikreligijny.pl/teksty/2062,witaj_pokarmie',
-              'http://spiewnikreligijny.pl/teksty/1096,witaj_zbawicielu',
-              'http://spiewnikreligijny.pl/teksty/460,witajmy_jezusa_nam_narodzonego',
-              'http://spiewnikreligijny.pl/teksty/471,witajze_dzieciatko',
-              'http://spiewnikreligijny.pl/teksty/237,witam_cie_witam_przenajswietsze_cialo',
-              'http://spiewnikreligijny.pl/teksty/465,wiwat_dzisiaj_boskiej_istnosci',
-              'http://spiewnikreligijny.pl/teksty/4829,woda_i_dzban_(kiedys_znow_bede_z_toba)',
-              'http://spiewnikreligijny.pl/teksty/472,wody_cedronu',
-              'http://spiewnikreligijny.pl/teksty/849,wody_jordanu',
-              'http://spiewnikreligijny.pl/teksty/2350,wody_nie_ugasza_jej',
-              'http://spiewnikreligijny.pl/teksty/3417,wojna', 'http://spiewnikreligijny.pl/teksty/3233,wojna_jest_zla',
-              'http://spiewnikreligijny.pl/teksty/2351,wojownicy_pana',
-              'http://spiewnikreligijny.pl/teksty/2035,wolanie_i', 'http://spiewnikreligijny.pl/teksty/1811,wolanie_ii',
-              'http://spiewnikreligijny.pl/teksty/1109,wolek_i_osiolek',
-              'http://spiewnikreligijny.pl/teksty/521,won_chrystusa',
-              'http://spiewnikreligijny.pl/teksty/3211,wprowadz_mnie_w_komnaty_twoje',
-              'http://spiewnikreligijny.pl/teksty/1223,wspaniala_matka',
-              'http://spiewnikreligijny.pl/teksty/850,wspanialy_dawco_milosci',
-              'http://spiewnikreligijny.pl/teksty/9,wstal_pan_chrystus_z_martwych',
-              'http://spiewnikreligijny.pl/teksty/3160,wstane_po_miescie_chodzic_bede',
-              'http://spiewnikreligijny.pl/teksty/470,wstancie_wstancie_wstancie_pasterze_mili',
-              'http://spiewnikreligijny.pl/teksty/3275,wszystkie_dzieci_ida_do_nieba',
-              'http://spiewnikreligijny.pl/teksty/851,wszystkie_moje_troski_i_klopoty',
-              'http://spiewnikreligijny.pl/teksty/1194,wszystkie_narody',
-              'http://spiewnikreligijny.pl/teksty/2352,wszystkie_nasze_dzienne_sprawy',
-              'http://spiewnikreligijny.pl/teksty/2353,wszystko_moge',
-              'http://spiewnikreligijny.pl/teksty/560,wszystko_moje_jest_twoje',
-              'http://spiewnikreligijny.pl/teksty/3125,wszystko_panie_ciebie_chwali',
-              'http://spiewnikreligijny.pl/teksty/1543,wszystko_to_co_mamy',
-              'http://spiewnikreligijny.pl/teksty/1657,wszystko_tobie_oddac_pragne',
-              'http://spiewnikreligijny.pl/teksty/461,wsrod_nocnej_ciszy',
-              'http://spiewnikreligijny.pl/teksty/4731,wuchta_wiary',
-              'http://spiewnikreligijny.pl/teksty/1552,wy_jestescie_na_ziemi',
-              'http://spiewnikreligijny.pl/teksty/2355,wy_jestescie_sola_ziemi',
-              'http://spiewnikreligijny.pl/teksty/852,wybacz_panie',
-              'http://spiewnikreligijny.pl/teksty/4853,wybieram_sie_stale',
-              'http://spiewnikreligijny.pl/teksty/239,wybral_nas_bog',
-              'http://spiewnikreligijny.pl/teksty/3312,wyciagam_do_ciebie_rece_swe',
-              'http://spiewnikreligijny.pl/teksty/3216,wyciagnij_rece',
-              'http://spiewnikreligijny.pl/teksty/2999,wydani_na_smierc',
-              'http://spiewnikreligijny.pl/teksty/3414,wyjdz',
-              'http://spiewnikreligijny.pl/teksty/2354,wykrzykujcie_bogu',
-              'http://spiewnikreligijny.pl/teksty/2457,wykrzykujcie_na_czesc_pana',
-              'http://spiewnikreligijny.pl/teksty/1003,wyplyn_na_glebie',
-              'http://spiewnikreligijny.pl/teksty/2356,wyruszyl_nasz_pan',
-              'http://spiewnikreligijny.pl/teksty/25,wyslawiajmy_chrysta_pana',
-              'http://spiewnikreligijny.pl/teksty/250,wysluchaj_panie', 'http://spiewnikreligijny.pl/teksty/928,wyspa',
-              'http://spiewnikreligijny.pl/teksty/4879,wystarczy',
-              'http://spiewnikreligijny.pl/teksty/853,wystarczy_bys_byl',
-              'http://spiewnikreligijny.pl/teksty/1940,wywyzszony', 'http://spiewnikreligijny.pl/teksty/854,wyznanie',
-              'http://spiewnikreligijny.pl/teksty/2067,wzniesmy_dlonie',
-              'http://spiewnikreligijny.pl/teksty/1279,wznosze_oczy_wysoko_do_wzgorz',
-              'http://spiewnikreligijny.pl/teksty/2194,wznosze_rece_me_wzwyz',
-              'http://spiewnikreligijny.pl/teksty/1514,wzywam_cie',
-              'http://spiewnikreligijny.pl/teksty/3793,z_aniolami',
-              'http://spiewnikreligijny.pl/teksty/2273,z_apostolem_radosci',
-              'http://spiewnikreligijny.pl/teksty/3176,z_calego_serca_chce_chwalic_pana',
-              'http://spiewnikreligijny.pl/teksty/4045,z_dalekiej_fatimy',
-              'http://spiewnikreligijny.pl/teksty/859,z_dawna_polski_tys_krolowa',
-              'http://spiewnikreligijny.pl/teksty/478,z_glebokosci_mego_grzechu',
-              'http://spiewnikreligijny.pl/teksty/477,z_gor_zejdziemy',
-              'http://spiewnikreligijny.pl/teksty/546,z_milosci_do_ciebie',
-              'http://spiewnikreligijny.pl/teksty/3208,z_milosci_umarl_pan',
-              'http://spiewnikreligijny.pl/teksty/475,z_narodzenia_pana',
-              'http://spiewnikreligijny.pl/teksty/476,z_nieba_wysokiego',
-              'http://spiewnikreligijny.pl/teksty/2946,z_rak_kaplanskich',
-              'http://spiewnikreligijny.pl/teksty/4132,z_rozancem_w_reku',
-              'http://spiewnikreligijny.pl/teksty/2884,z_skalnych_portugalskich_gor',
-              'http://spiewnikreligijny.pl/teksty/590,z_ta_sama_miloscia_w_tym_samym_duchu',
-              'http://spiewnikreligijny.pl/teksty/243,z_tej_biednej_ziemi',
-              'http://spiewnikreligijny.pl/teksty/861,z_toba_ciemnosc',
-              'http://spiewnikreligijny.pl/teksty/246,z_twojego_boku_chryste',
-              'http://spiewnikreligijny.pl/teksty/2157,z_wiara_w_sercu',
-              'http://spiewnikreligijny.pl/teksty/1978,z_wielka_ufnoscia',
-              'http://spiewnikreligijny.pl/teksty/247,za_cozes_ty_jezu',
-              'http://spiewnikreligijny.pl/teksty/3260,za_malo',
-              'http://spiewnikreligijny.pl/teksty/902,za_malo_kochalem_cie_panie_(hymn_o_milosci)',
-              'http://spiewnikreligijny.pl/teksty/856,za_nami_wstawiaj_sie_maryjo',
-              'http://spiewnikreligijny.pl/teksty/3967,za_ten_posilek',
-              'http://spiewnikreligijny.pl/teksty/4830,za_wolnosc',
-              'http://spiewnikreligijny.pl/teksty/3263,za_zu_za_zi', 'http://spiewnikreligijny.pl/teksty/2151,zadumka',
-              'http://spiewnikreligijny.pl/teksty/530,zagubienie',
-              'http://spiewnikreligijny.pl/teksty/248,zagubiony_slad',
-              'http://spiewnikreligijny.pl/teksty/480,zajasniala_cudna_zorza',
-              'http://spiewnikreligijny.pl/teksty/2173,zakochaj_sie_w_panu',
-              'http://spiewnikreligijny.pl/teksty/855,zakwitla_noc',
-              'http://spiewnikreligijny.pl/teksty/1908,zamieniles_w_taniec',
-              'http://spiewnikreligijny.pl/teksty/1080,zamknij_swe_powieczki_',
-              'http://spiewnikreligijny.pl/teksty/857,zapada_zmrok',
-              'http://spiewnikreligijny.pl/teksty/1734,zapada_zmrok_nad_watykanem',
-              'http://spiewnikreligijny.pl/teksty/2952,zaproscie_jezusa',
-              'http://spiewnikreligijny.pl/teksty/483,zasypia_swiat_mrozna_noca',
-              'http://spiewnikreligijny.pl/teksty/3560,zasnij_dziecino',
-              'http://spiewnikreligijny.pl/teksty/1047,zaspiewac_chce',
-              'http://spiewnikreligijny.pl/teksty/2357,zaspiewajcie_panu',
-              'http://spiewnikreligijny.pl/teksty/3154,zaspiewajcie_serca_nasze',
-              'http://spiewnikreligijny.pl/teksty/482,zaspiewajmy_kolede_jezusowi',
-              'http://spiewnikreligijny.pl/teksty/481,zaspiewajmy_wszyscy_razem',
-              'http://spiewnikreligijny.pl/teksty/3193,zaspiewam_bogu',
-              'http://spiewnikreligijny.pl/teksty/1013,zatanczmy_panu',
-              'http://spiewnikreligijny.pl/teksty/1541,zaufaj_panu_juz_dzis',
-              'http://spiewnikreligijny.pl/teksty/3525,zaufalem_panu',
-              'http://spiewnikreligijny.pl/teksty/2958,zaufanie',
-              'http://spiewnikreligijny.pl/teksty/3509,zawieja_i_beznadzieja',
-              'http://spiewnikreligijny.pl/teksty/651,zawitaj_krolowo_rozanca_swietego',
-              'http://spiewnikreligijny.pl/teksty/3371,zawitaj_ukrzyzowany',
-              'http://spiewnikreligijny.pl/teksty/240,zawitaj_ukrzyzowany',
-              'http://spiewnikreligijny.pl/teksty/2993,zawrocmy_do_pana',
-              'http://spiewnikreligijny.pl/teksty/1126,zawsze_z_toba_byc_',
-              'http://spiewnikreligijny.pl/teksty/3028,zbawienie_boze',
-              'http://spiewnikreligijny.pl/teksty/2206,zbawienie_jest_w_bogu_mym',
-              'http://spiewnikreligijny.pl/teksty/241,zbawienie_przyszlo_przez_krzyz',
-              'http://spiewnikreligijny.pl/teksty/244,zblizam_sie_ktobie',
-              'http://spiewnikreligijny.pl/teksty/858,zbuduj_ojsze_nasz',
-              'http://spiewnikreligijny.pl/teksty/3217,zbytnie_troski',
-              'http://spiewnikreligijny.pl/teksty/50,zdrowas_badz_maryja',
-              'http://spiewnikreligijny.pl/teksty/1022,zdrowas_maryjo_bogarodzico',
-              'http://spiewnikreligijny.pl/teksty/3243,zgadzam_sie', 'http://spiewnikreligijny.pl/teksty/535,ziarno',
-              'http://spiewnikreligijny.pl/teksty/1743,ziemia_ktora_mi_dajesz',
-              'http://spiewnikreligijny.pl/teksty/72,ziemia_swieta',
-              'http://spiewnikreligijny.pl/teksty/3725,ziemio_raduj_sie',
-              'http://spiewnikreligijny.pl/teksty/3330,zion_(psalm_13)',
-              'http://spiewnikreligijny.pl/teksty/479,zjawilo_sie_nam_dzis_cos_nowego',
-              'http://spiewnikreligijny.pl/teksty/1874,zjednoczeni_w_duchu',
-              'http://spiewnikreligijny.pl/teksty/860,zlo_dobrem_zwyciezaj',
-              'http://spiewnikreligijny.pl/teksty/328,zlota_jerozolima_i_biedne_betlejem',
-              'http://spiewnikreligijny.pl/teksty/1280,zlozylem_w_panu_cala_nadzieje',
-              'http://spiewnikreligijny.pl/teksty/29,zlozcie_troski_zalujacy',
-              'http://spiewnikreligijny.pl/teksty/3729,zmartwychwstal_',
-              'http://spiewnikreligijny.pl/teksty/3115,zmartwychwstal_(zatryumfowal_w_niebie)',
-              'http://spiewnikreligijny.pl/teksty/17,zmartwychwstal_jezus_nasz_pan',
-              'http://spiewnikreligijny.pl/teksty/19,zmartwychwstal_pan_(gdzie_jestes_o_smierci)',
-              'http://spiewnikreligijny.pl/teksty/18,zmartwychwstal_pan_i_zyje_dzis',
-              'http://spiewnikreligijny.pl/teksty/16,zmartwychwstal_pan_zwycieska_chwala_niebo_brzmi',
-              'http://spiewnikreligijny.pl/teksty/15,zmartwychwstaly_pan',
-              'http://spiewnikreligijny.pl/teksty/242,zmiluj_sie_nade_mna_boze',
-              'http://spiewnikreligijny.pl/teksty/2939,znak_krzyza',
-              'http://spiewnikreligijny.pl/teksty/2313,znak_pokoju',
-              'http://spiewnikreligijny.pl/teksty/249,znak_tej_ziemi',
-              'http://spiewnikreligijny.pl/teksty/3061,znalazles_mnie_panie',
-              'http://spiewnikreligijny.pl/teksty/3239,znowu_cie_spotykam',
-              'http://spiewnikreligijny.pl/teksty/3235,zobaczcie_jak_jest_pieknie',
-              'http://spiewnikreligijny.pl/teksty/2203,zobaczcie_jak_wielka_milosc',
-              'http://spiewnikreligijny.pl/teksty/2963,zobaczcie_jaka_miloscia',
-              'http://spiewnikreligijny.pl/teksty/3014,zostan_misjonarzem',
-              'http://spiewnikreligijny.pl/teksty/3850,zostan_panem_mym',
-              'http://spiewnikreligijny.pl/teksty/2358,zostan_tu',
-              'http://spiewnikreligijny.pl/teksty/2376,zostan_z_nami',
-              'http://spiewnikreligijny.pl/teksty/2315,zrzuc_ciezary_na_jezusa_(wyzej_wyzej)',
-              'http://spiewnikreligijny.pl/teksty/56,zwiastowanie',
-              'http://spiewnikreligijny.pl/teksty/3583,zwiastunom_z_gor',
-              'http://spiewnikreligijny.pl/teksty/8,zwyciezca_smierci',
-              'http://spiewnikreligijny.pl/teksty/3000,zrodla', 'http://spiewnikreligijny.pl/teksty/4770,zrodlo',
-              'http://spiewnikreligijny.pl/teksty/3001,zadna_piesn',
-              'http://spiewnikreligijny.pl/teksty/4833,zycie_na_szkode',
-              'http://spiewnikreligijny.pl/teksty/4754,zyczenia',
-              'http://spiewnikreligijny.pl/teksty/3073,zycze_ci_przyjaciela',
-              'http://spiewnikreligijny.pl/teksty/3199,zyc_w_spolecznosci_z_toba',
-              'http://spiewnikreligijny.pl/teksty/2092,zyc_zyc_kochac']
+links_list = ['http://spiewnikreligijny.pl/lista-tekstow/']
